Remove D2CommonDefinitions/DataTables.h and start moving structures to the correct headers
