+++ source/D2Common/include/D2DataTbls.h
@@ -1,8 +1,470 @@
 #pragma once
 
-#include "CommonDefinitions.h"
-#include "D2Math.h"
-#include <cstdio>
+#include <D2BasicTypes.h>
+#include <Fog.h>
+#include <stdio.h> // for FILE
+
+#include "DataTbls/AnimTbls.h"
+#include "DataTbls/ArenaTbls.h"
+#include "DataTbls/BeltsTbls.h"
+#include "DataTbls/DataTbls.h"
+#include "DataTbls/FieldTbls.h"
+#include "DataTbls/HoradricCube.h"
+#include "DataTbls/InvTbls.h"
+#include "DataTbls/ItemsTbls.h"
+#include "DataTbls/LevelsTbls.h"
+#include "DataTbls/MissilesTbls.h"
+#include "DataTbls/MonsterTbls.h"
+#include "DataTbls/ObjectsTbls.h"
+#include "DataTbls/OverlayTbls.h"
+#include "DataTbls/SkillsTbls.h"
+#include "DataTbls/SequenceTbls.h"
+#include "DataTbls/TokenTbls.h"
+#include "DataTbls/TransformTbls.h"
+#include <D2Seed.h>
+
+
+
+struct D2BodyLocsTxt
+{
+	uint32_t dwCode;						//0x00
+};
+
+
+struct D2CharStatsTxt
+{
+	wchar_t wszClassName[16];				//0x00
+	char szClassName[16];					//0x20
+	uint8_t nStr;							//0x30
+	uint8_t nDex;							//0x31
+	uint8_t nInt;							//0x32
+	uint8_t nVit;							//0x33
+	uint8_t nStamina;						//0x34
+	uint8_t nLifeAdd;						//0x35
+	uint8_t nPercentStr;					//0x36
+	uint8_t nPercentInt;					//0x37
+	uint8_t nPercentDex;					//0x38
+	uint8_t nPercentVit;					//0x39
+	uint8_t nManaRegen;						//0x3A
+	uint8_t pad0x3B;						//0x3B
+	uint32_t dwToHitFactor;					//0x3C
+	uint8_t nWalkSpeed;						//0x40
+	uint8_t nRunSpeed;						//0x41
+	uint8_t nRunDrain;						//0x42
+	uint8_t nLifePerLevel;					//0x43
+	uint8_t nStaminaPerLevel;				//0x44
+	uint8_t nManaPerLevel;					//0x45
+	uint8_t nLifePerVitality;				//0x46
+	uint8_t nStaminaPerVitality;			//0x47
+	uint8_t nManaPerMagic;					//0x48
+	uint8_t nBlockFactor;					//0x49
+	uint16_t pad0x4A;						//0x4A
+	uint32_t dwBaseWClass;					//0x4C
+	uint8_t nStatsPerLevel;					//0x50
+	uint8_t pad0x51;						//0x51
+	uint16_t wStrAllSkills;					//0x52
+	uint16_t wStrSkillTab[3];				//0x54
+	uint16_t wStrClassOnly;					//0x5A
+	D2CharItemStrc pItems[10];				//0x5C
+	uint16_t wStartSkill;					//0xAC
+	int16_t nBaseSkill[10];					//0xAE
+	uint16_t pad0xC2;						//0xC2
+};
+
+struct D2DifficultyLevelsTxt
+{
+	uint32_t dwResistPenalty;				//0x00
+	uint32_t dwDeathExpPenalty;				//0x04
+	uint32_t dwUberCodeOddsNorm;			//0x08
+	uint32_t dwUberCodeOddsGood;			//0x0C
+	uint32_t dwMonsterSkillBonus;			//0x10
+	uint32_t dwMonsterFreezeDiv;			//0x14
+	uint32_t dwMonsterColdDiv;				//0x18
+	uint32_t dwAiCurseDiv;					//0x1C
+	uint32_t dwUltraCodeOddsNorm;			//0x20
+	uint32_t dwUltraCodeOddsGood;			//0x24
+	uint32_t dwLifeStealDiv;				//0x28
+	uint32_t dwManaStealDiv;				//0x2C
+	uint32_t dwUniqueDmgBonus;				//0x30
+	uint32_t dwChampionDmgBonus;			//0x34
+	uint32_t dwHireableBossDmgPercent;		//0x38
+	uint32_t dwMonsterCEDmgPercent;			//0x3C
+	uint32_t dwStaticFieldMin;				//0x40
+	uint32_t dwGambleRare;					//0x44
+	uint32_t dwGambleSet;					//0x48
+	uint32_t dwGambleUniq;					//0x4C
+	uint32_t dwGambleUber;					//0x50
+	uint32_t dwGambleUltra;					//0x54
+};
+
+struct D2HireDescTxt
+{
+	uint32_t dwCode;						//0x00
+};
+
+struct D2HitClassTxt
+{
+	uint32_t dwCode;						//0x00
+};
+
+struct D2MissCalcTxt
+{
+	uint32_t dwCode;						//0x00
+};
+
+struct D2EventsTxt
+{
+	uint16_t wEvent;						//0x00
+};
+
+struct D2ElemTypesTxt
+{
+	uint32_t dwCode;						//0x00
+};
+
+struct D2ColorsTxt
+{
+	uint32_t dwCode;						//0x00
+};
+
+struct D2ExperienceTxt
+{
+	uint32_t dwClass[7];					//0x00
+	uint32_t dwExpRatio;					//0x1C
+};
+
+struct D2PlayerClassTxt
+{
+	uint32_t dwCode;						//0x00
+};
+
+struct D2PlrModeTxtStub
+{
+	uint32_t dwCode;						//0x00
+};
+
+struct D2MonAiTxt
+{
+	uint16_t wAi;							//0x00
+};
+
+struct D2MonModeTxtStub
+{
+	uint32_t dwCode;						//0x00
+};
+
+struct D2MonPlaceTxt
+{
+	uint16_t nCode;							//0x00
+};
+
+struct D2StorePageTxt
+{
+	uint32_t dwCode;						//0x00
+};
+
+struct D2PetTypeTxt
+{
+	uint32_t dwHcIdx;						//0x00
+	uint32_t dwPetFlags;					//0x04
+	uint16_t wGroup;						//0x08
+	uint16_t wBaseMax;						//0x0A
+	uint16_t wStringId;						//0x0C
+	uint8_t nIconType;						//0x0E
+	char szBaseicon[32];					//0x0F
+	char szIcon[4][32];						//0x2F
+	uint8_t pad0xAF[3];						//0xAF
+	uint16_t wClass[4];						//0xB2
+	uint8_t unk0xBA[2];						//0xBA
+	int32_t nSkillCount;					//0xBC
+	uint16_t wSkillIds[15];					//0xC0
+	uint16_t unk0xDE;						//0xDE
+};
+
+struct D2StatesTxt
+{
+	uint16_t wState;						//0x00
+	uint16_t wOverlay[4];					//0x02
+	uint16_t wCastOverlay;					//0x0A
+	uint16_t wRemoveOverlay;				//0x0C
+	uint16_t wPrgOverlay;					//0x0E
+	union
+	{
+		uint8_t nStateFlags[4];				//0x10
+		uint32_t dwStateFlags;				//0x10
+	};
+	uint32_t dwStateFlagsEx;				//0x14
+	uint16_t wStat;							//0x18
+	uint16_t wSetFunc;						//0x1A
+	uint16_t wRemFunc;						//0x1C
+	uint16_t wGroup;						//0x1E
+	uint8_t nColorPri;						//0x20
+	uint8_t nColorShift;					//0x21
+	uint8_t nLightRGB[3];					//0x22
+	uint8_t pad0x25;						//0x25
+	uint16_t wOnSound;						//0x26
+	uint16_t wOffSound;						//0x28
+	uint16_t wItemType;						//0x2A
+	char nItemTrans;						//0x2C
+	uint8_t nGfxType;						//0x2D
+	uint16_t wGfxClass;						//0x2E
+	uint16_t wCltEvent;						//0x30
+	uint16_t wCltEventFunc;					//0x32
+	uint16_t wCltActiveFunc;				//0x34
+	uint16_t wSrvActiveFunc;				//0x36
+	uint16_t wSkill;						//0x38
+	uint16_t wMissile;						//0x3A
+};
+
+//TODO: see comments
+struct D2DataTablesStrc
+{
+	D2PlayerClassTxt* pPlayerClassTxt;					//0x00
+	D2TxtLinkStrc* pPlayerClassLinker;					//0x04
+	D2BodyLocsTxt* pBodyLocsTxt;						//0x08
+	D2TxtLinkStrc* pBodyLocsLinker;						//0x0C
+	D2StorePageTxt* pStorePageTxt;						//0x10
+	D2TxtLinkStrc* pStorePageLinker;					//0x14
+	D2ElemTypesTxt* pElemTypesTxt;						//0x18
+	D2TxtLinkStrc* pElemTypesLinker;					//0x1C
+	D2HitClassTxt* pHitClassTxt;						//0x20
+	D2TxtLinkStrc* pHitClassLinker;						//0x24
+	D2MonModeTxtStub* pMonModeTxtStub;					//0x28
+	D2TxtLinkStrc* pMonModeLinker;						//0x2C
+	D2PlrModeTxtStub* pPlrModeTxtStub;					//0x30
+	D2TxtLinkStrc* pPlrModeLinker;						//0x34
+	D2SkillCalcTxt* pSkillCalcTxt;						//0x38
+	D2TxtLinkStrc* pSkillCalcLinker;					//0x3C
+	char* pSkillsCode;									//0x40
+	unsigned int nSkillsCodeSize;						//0x44
+	unsigned int nSkillsCodeSizeEx;						//0x48
+	char* pSkillDescCode;								//0x4C
+	unsigned int nSkillDescCodeSize;					//0x50
+	unsigned int nSkillDescCodeSizeEx;					//0x54
+	D2MissCalcTxt* pMissileCalcTxt;						//0x58
+	D2TxtLinkStrc* pMissileCalcLinker;					//0x5C
+	char* pMissCode;									//0x60
+	unsigned int nMissCodeSize;							//0x64
+	int nMissCodeSizeEx;								//0x68
+
+	/**/	const char* pSkillCode;                             //+0000006C   052C445C      skillcode.txt (Id from skills.txt)
+	/**/	D2TxtLinkStrc* iSkillCode;                       //+00000070   0141F084      skillcode info
+
+	D2EventsTxt* pEventsTxt;							//0x74
+	D2TxtLinkStrc* pEventsLinker;						//0x78
+	D2CompCodeTxt* pCompCodeTxt;						//0x7C
+	D2TxtLinkStrc* pCompCodeLinker;						//0x80
+	int nCompCodeTxtRecordCount;						//0x84
+	D2MonAiTxt* pMonAiTxt;								//0x88
+	D2TxtLinkStrc* pMonAiLinker;						//0x8C
+	int nMonAiTxtRecordCount;							//0x90
+	D2TxtLinkStrc* pItemsLinker;						//0x94
+	char* pItemsCode;									//0x98
+	unsigned int nItemsCodeSize;						//0x9C
+	int nItemsCodeSizeEx;								//0xA0
+	D2PropertiesTxt* pPropertiesTxt;					//0xA4
+	D2TxtLinkStrc* pPropertiesLinker;					//0xA8
+	int nPropertiesTxtRecordCount;						//0xAC
+	D2TxtLinkStrc* pRunesLinker;						//0xB0
+	D2HireDescTxt* pHireDescTxt;						//0xB4
+	D2TxtLinkStrc* pHireDescLinker;						//0xB8
+	D2StatesTxt* pStatesTxt;							//0xBC
+	D2TxtLinkStrc* pStatesLinker;						//0xC0
+	int nStatesTxtRecordCount;							//0xC4
+	uint32_t* pStateMasks;								//0xC8
+	uint32_t* fStateMasks[40];							//0xCC
+	short* pProgressiveStates;							//0x16C
+	int nProgressiveStates;								//0x170
+	short* pCurseStates;								//0x174
+	int nCurseStates;									//0x178
+	short* pTransformStates;							//0x17C
+	int nTransformStates;								//0x180
+	short* pActionStates;								//0x184
+	int nActionStates;									//0x188
+	short* pColourStates;								//0x18C
+	int nColourStates;									//0x190
+	D2SoundsTxtStub* pSoundsTxtCodes;					//0x194
+	D2TxtLinkStrc* pSoundsLinker;						//0x198
+	int nSoundsTxtCodes;								//0x19C
+	D2HirelingTxt* pHirelingTxt;						//0x1A0
+	int nHirelingTxtRecordCount;						//0x1A4
+	int nClassicHirelingStartRecordIds[256];			//0x1A8
+	int nExpansionHirelingStartRecordIds[256];			//0x5A8
+	D2NpcTxt* pNpcTxt;									//0x9A8
+	int nNpcTxtRecordCount;								//0x9AC
+	D2ColorsTxt* pColorsTxt;							//0x9B0
+	D2TxtLinkStrc* pColorsLinker;						//0x9B4
+	D2TxtLinkStrc* pTreasureClassExLinker;				//0x9B8
+	D2TCExShortStrc* pTreasureClassEx;					//0x9BC
+	int nTreasureClassEx;								//0x9C0
+	D2TCExShortStrc* pChestTreasureClasses[45];			//0x9C4
+	D2MonStatsTxt* pMonStatsTxt;						//0xA78
+	D2TxtLinkStrc* pMonStatsLinker;						//0xA7C
+	int nMonStatsTxtRecordCount;						//0xA80
+	D2MonSoundsTxt* pMonSoundsTxt;						//0xA84
+	D2TxtLinkStrc* pMonSoundsLinker;					//0xA88
+	int nMonSoundsTxtRecordCount;						//0xA8C
+	D2MonStats2Txt* pMonStats2Txt;						//0xA90
+	D2TxtLinkStrc* pMonStats2Linker;					//0xA94
+	int nMonStats2TxtRecordCount;						//0xA98
+	D2MonPlaceTxt* pMonPlaceTxt;						//0xA9C
+	D2TxtLinkStrc* pMonPlaceLinker;						//0xAA0
+	int nMonPlaceTxtRecordCount;						//0xAA4
+	D2MonPresetTxt* pMonPresetTxt;						//0xAA8
+	D2MonPresetTxt* pMonPresetTxtActSections[5];		//0xAAC
+	int nMonPresetTxtActRecordCounts[5];				//0xAC0
+	D2SuperUniquesTxt* pSuperUniquesTxt;				//0xAD4
+	D2TxtLinkStrc* pSuperUniquesLinker;					//0xAD8
+	int nSuperUniquesTxtRecordCount;					//0xADC
+	short nSuperUniqueIds[66];							//0xAE0
+	D2MissilesTxt* pMissilesTxt;						//0xB64
+	D2TxtLinkStrc* pMissilesLinker;						//0xB68
+	int nMissilesTxtRecordCount;						//0xB6C
+	D2MonLvlTxt* pMonLvlTxt;							//0xB70
+	int nMonLvlTxtRecordCount;							//0xB74
+	D2MonSeqTxt* pMonSeqTxt;							//0xB78
+	D2TxtLinkStrc* pMonSeqLinker;						//0xB7C
+	int nMonSeqTxtRecordCount;							//0xB80
+	D2SeqRecordStrc* pMonSeqTable;						//0xB84
+	int nMonSeqTableRecordCount;						//0xB88
+	D2SkillDescTxt* pSkillDescTxt;						//0xB8C
+	D2TxtLinkStrc* pSkillDescLinker;					//0xB90
+	int nSkillDescTxtRecordCount;						//0xB94
+	D2SkillsTxt* pSkillsTxt;							//0xB98
+	D2TxtLinkStrc* pSkillsLinker;						//0xB9C
+	int nSkillsTxtRecordCount;							//0xBA0
+	int* nClassSkillCount;								//0xBA4
+	int nHighestClassSkillCount;						//0xBA8
+	short* nClassSkillList;								//0xBAC
+	int nPassiveSkills;									//0xBB0
+	uint16_t* pPassiveSkills;							//0xBB4
+	D2TxtLinkStrc* pOverlayLinker;						//0xBB8
+	D2OverlayTxt* pOverlayTxt;							//0xBBC
+	int nOverlayTxtRecordCount;							//0xBC0
+	D2CharStatsTxt* pCharStatsTxt;						//0xBC4
+	int nCharStatsTxtRecordCount;						//0xBC8
+	D2ItemStatCostTxt* pItemStatCostTxt;				//0xBCC
+	D2TxtLinkStrc* pItemStatCostLinker;					//0xBD0
+	int nItemStatCostTxtRecordCount;					//0xBD4
+	uint16_t* pStatsWithDescFunc;						//0xBD8
+	int nStatsWithDescFunc;								//0xBDC
+	D2MonEquipTxt* pMonEquipTxt;						//0xBE0
+	int nMonEquipTxtRecordCount;						//0xBE4
+	D2PetTypeTxt* pPetTypeTxt;							//0xBE8
+	D2TxtLinkStrc* pPetTypeLinker;						//0xBEC
+	int nPetTypeTxtRecordCount;							//0xBF0
+	D2TxtLinkStrc* pItemTypesLinker;					//0xBF4
+	D2ItemTypesTxt* pItemTypesTxt;						//0xBF8
+	int nItemTypesTxtRecordCount;						//0xBFC
+	int nItemTypesIndex;								//0xC00
+	uint32_t* pItemTypesNest;							//0xC04
+	D2TxtLinkStrc* pSetsLinker;							//0xC08
+	D2SetsTxt* pSetsTxt;								//0xC0C
+	int nSetsTxtRecordCount;							//0xC10
+	D2TxtLinkStrc* pSetItemsLinker;						//0xC14
+	D2SetItemsTxt* pSetItemsTxt;						//0xC18
+	int nSetItemsTxtRecordCount;						//0xC1C
+	D2TxtLinkStrc* pUniqueItemsLinker;					//0xC20
+	D2UniqueItemsTxt* pUniqueItemsTxt;					//0xC24
+	int nUniqueItemsTxtRecordCount;						//0xC28
+	D2TxtLinkStrc* pMonPropLinker;						//0xC2C
+	D2MonPropTxt* pMonPropTxt;							//0xC30
+	int nMonPropTxtRecordCount;							//0xC34
+	D2TxtLinkStrc* pMonTypeLinker;						//0xC38
+	D2MonTypeTxt* pMonTypeTxt;							//0xC3C
+	int nMonTypeTxtRecordCount;							//0xC40
+	uint32_t* pMonTypeNest;								//0xC44
+	int nMonTypeIndex;									//0xC48
+	D2TxtLinkStrc* pMonUModLinker;						//0xC4C
+	D2MonUModTxt* pMonUModTxt;							//0xC50
+	int nMonUModTxtRecordCount;							//0xC54
+	D2LevelsTxt* pLevelsTxt;							//0xC58
+	int nLevelsTxtRecordCount;							//0xC5C
+	D2LevelDefBin* pLevelDefBin;						//0xC60
+	D2LvlPrestTxt* pLvlPrestTxt;						//0xC64
+	int nLvlPrestTxtRecordCount;						//0xC68
+	int nStuff;											//0xC6C
+	int nShiftedStuff;									//0xC70
+	D2AnimDataStrc* pAnimData;							//0xC74
+	D2ExperienceTxt* pExperienceTxt;					//0xC78
+	D2DifficultyLevelsTxt* pDifficultyLevelsTxt;		//0xC7C
+	int nDifficultyLevelsTxtRecordCount;				//0xC80
+	BOOL bCompileTxt;									//0xC84
+	int ExpFieldI[9];									//0xC88
+	char* pFieldData;									//0xCAC
+	D2CollisionFieldStrc pCollisionField;				//0xCB0
+	D2CubeMainTxt* pCubeMainTxt;						//0xCC4
+	int nCubeMainTxtRecordCount;						//0xCC8
+	int nInventoryTxtRecordCount;						//0xCCC
+	D2InventoryTxt* pInventoryTxt;						//0xCD0
+
+	uint32_t unk0xCD4;									//0xCD4
+
+	D2ItemDataTbl pItemDataTables;						//0xCD8
+	uint16_t* pIndexOldToCurrent;						//0xCF8	
+	D2GemDataTbl pGemDataTables;						//0xCFC
+	D2LowQualityItemDataTbl pLowQualityItemDataTables;	//0xD04
+	D2BookDataTbl pBookDataTables;						//0xD0C
+	D2RareAffixDataTbl pRareAffixDataTables;			//0xD14
+	D2ItemRatioDataTbl pItemRatioDataTables;			//0xD24
+
+	uint32_t unk0xD2C;									//0xD2C
+
+	D2GambleDataTbl pGambleDataTables;					//0xD30
+	D2MagicAffixDataTbl pMagicAffixDataTables;			//0xEC8
+	D2RuneDataTbl pRuneDataTables;						//0xEDC
+	D2QualityItemDataTbl pQualityItemDataTables;		//0xEE4
+
+	uint32_t unk0xEEC;									//0xEEC
+
+	D2SeedStrc pAutomapSeed;							//0xEF0
+	D2AutomapRandStrc pAutoMapRand[36];					//0xEF8
+	D2LvlTypesTxt* pLvlTypesTxt;						//0x1018
+	int* pPortalLevels;									//0x101C
+	int nPortalLevels;									//0x1020
+	int nLvlTypesTxtRecordCount;						//0x1024
+	D2LvlWarpTxt* pLvlWarpTxt;							//0x1028
+	int nLvlWarpTxtRecordCount;							//0x102C
+	D2LvlMazeTxt* pLvlMazeTxt;							//0x1030
+	int nLvlMazeTxtRecordCount;							//0x1034
+	D2LvlSubTxt* pLvlSubTxt;							//0x1038
+	int nLvlSubTxtRecordCount;							//0x103C
+	int* pLvlSubTypeStartIds;							//0x1040
+	D2TileLibraryHashStrc** ppTileLibraryHash;			//0x1044
+	D2DrlgFileStrc** ppLvlPrestFiles;					//0x1048
+	D2DrlgFileStrc** ppLvlSubTypeFiles;					//0x104C
+	D2AutomapShortStrc* pAutoMap;						//0x1050
+	int nAutoMap;										//0x1054
+	char szDefaultString;								//0x1058
+	char pad0x1059[3];									//0x1059
+	D2MonItemPercentDataTbl pMonItemPercentDataTables;	//0x105C
+	wchar_t wszDefault;									//0x1064
+	uint16_t unk0x1066;									//0x1066
+	D2UniqueTitleTxt* pUniqueTitleTxt;					//0x1068
+	D2UniquePrefixTxt* pUniquePrefixTxt;				//0x106C
+	D2UniqueSuffixTxt* pUniqueSuffixTxt;				//0x1070
+	D2UniqueAppellationTxt* pUniqueAppellationTxt;		//0x1074
+	int nUniqueTitleTxtRecordCount;						//0x1078
+	int nUniquePrefixTxtRecordCount;					//0x107C
+	int nUniqueSuffixTxtRecordCount;					//0x1080
+	int nUniqueAppellationTxtRecordCount;				//0x1084
+
+	uint32_t unk7[2];									//+00001088
+	short nTreasureClassItemTypes;						//0x1090
+	short unk0x1092;									//0x1092
+
+	D2ShrinesTxt* pShrinesTxt;							//0x1094
+	int nShrinesTxtRecordCount;							//0x1098
+	D2ObjectsTxt* pObjectsTxt;							//0x109C
+	int nObjectsTxtRecordCount;							//0x10A0
+	D2ObjGroupTxt* pObjGroupTxt;						//0x10A4
+	int nObjGroupTxtRecordCount;						//0x10A8
+	D2ArmTypeTxt* pArmTypeTxt;							//0x10AC
+	D2MonModeDataTbl pMonModeDataTables;				//0x10B0
+	D2CompositTxt* pCompositTxt;						//0x10C0
+	D2ObjModeDataTbl pObjModeDataTables;				//0x10C4
+	D2PlrModeDataTbl pPlrModeDataTables;				//0x10D4
+};
 
 
 extern D2ArenaTxt* gpArenaTxtTable;
@@ -37,46 +499,6 @@ void* __fastcall DATATBLS_GetBinaryData(void* pMemPool, char* szFileName, int* p
 
 
 
-//D2Common.0x6FD473C0
-D2AnimDataStrc* __fastcall DATATBLS_LoadAnimDataD2(void* pMemPool);
-//D2Common.0x6FD47460
-void __fastcall DATATBLS_UnloadAnimDataD2(D2AnimDataStrc* pAnimData);
-//D2Common.0x6FD474A0
-D2AnimDataRecordStrc* __fastcall DATATBLS_GetAnimDataRecord(D2UnitStrc* pUnit, int nClassId, int nMode, int nUnitType, D2InventoryStrc* pInventory);
-//D2Common.0x6FD475D0 (#10640)
-void __stdcall D2Common_10640(D2UnitStrc* pUnit, int nAnimSpeed);
-//D2Common.0x6FD47700 (#10641)
-BOOL __stdcall D2Common_10641(char* szPath, int* pLength, int* pOutAnimSpeed, int* pOutFirstFrameTagged);
-
-
-
-//D2Common.0x6FD47840
-void __fastcall DATATBLS_LoadArenaTxt(void* pMemPool);
-//D2Common.0x6FD47970 (#10596)
-D2ArenaTxt* __fastcall DATATBLS_GetArenaTxtRecord(int nArenaType);
-//D2Common.0x6FD479B0
-void __fastcall DATATBLS_UnloadArenaTxt();
-//D2Common.0x6FD479D0
-void __fastcall DATATBLS_LoadCharTemplateTxt(void* pMemPool);
-//D2Common.0x6FD48770
-void __fastcall DATATBLS_UnloadCharTemplateTxt();
-//D2Common.0x6FD48790 (#10664)
-int __fastcall DATATBLS_GetCharTemplateTxtRecordCount();
-//D2Common.0x6FD487A0 (#10665)
-D2CharTemplateTxt* __fastcall DATATBLS_GetCharTemplateTxtRecord(int nTemplate, int nLevel);
-//D2Common.0x6FD48810 (#10666)
-int __fastcall DATATBLS_GetClassFromCharTemplateTxtRecord(int nTemplate, int nLevel);
-
-
-
-//D2Common.0x6FD48880
-void __fastcall DATATBLS_LoadBeltsTxt(void* pMemPool);
-//D2Common.0x6FD493A0
-void __fastcall DATATBLS_UnloadBeltsTxt();
-//D2Common.0x6FD493B0 (#10638)
-void __stdcall DATATBLS_GetBeltsTxtRecord(int nIndex, int bHigherRes, D2BeltsTxt* pRecord);
-//D2Common.0x6FD49420 (#10639)
-void __stdcall DATATBLS_GetInvRectFromBeltsTxt(int nIndex, int bHigherRes, D2InvRectStrc* pInvRect, int nBoxId);
 
 
 
@@ -99,26 +521,6 @@ void __fastcall DATATBLS_GetBinFileHandle(void* pMemPool, char* szFile, void** p
 //D2Common.0x6FD49850
 int __fastcall DATATBLS_AppendMemoryBuffer(char** ppCodes, int* pSize, int* pSizeEx, char* pBuffer, int nBufferSize);
 
-
-
-//D2Common.0x6FD498D0
-int __fastcall DATATBLS_MapSkillsTxtKeywordToNumber(char* szKey);
-//D2Common.0x6FD49980
-int __fastcall sub_6FD49980(int nValue);
-//D2Common.0x6FD49990
-int __fastcall sub_6FD49990(char* szText, int* a2, int a3, int nKeywordNumber);
-//D2Common.0x6FD49D20
-void __fastcall DATATBLS_SkillCalcLinker(char* pSrc, void* pRecord, int nOffset, int nPosition, int nTxtRow, int nTxtColumn);
-//D2Common.0x6FD49E40
-void __fastcall DATATBLS_SkillDescCalcLinker(char* pSrc, void* pRecord, int nOffset, int nPosition, int nTxtRow, int nTxtColumn);
-//D2Common.0x6FD49F60
-void __fastcall DATATBLS_LoadSkills_SkillDescTxt(void* pMemPool);
-//D2Common.0x6FD4E350
-void __fastcall DATATBLS_UnloadSkills_SkillDescTxt();
-//Inlined at various places
-D2SkillsTxt* __fastcall DATATBLS_GetSkillsTxtRecord(int nSkillId);
-//Inlined at various places
-D2SkillDescTxt* __fastcall DATATBLS_GetSkillDescTxtRecord(int nSkillDesc);
 
 
 
@@ -157,486 +559,3 @@ void __fastcall DATATBLS_LoadCharStatsTxt(void* pMemPool);
 //D2Common.0x6FD51BF0
 void __fastcall DATATBLS_LoadDifficultyLevelsTxt(void* pMemPool);
 
-
-
-//D2Common.0x6FD51FC0
-BOOL __fastcall DATATBLS_LoadExpFieldD2(void* pMemPool);
-//D2Common.0x6FD52010 (#11089)
-BOOL __stdcall DATATBLS_InitializeCollisionFieldTable(char* pExpField, int nSize);
-//D2Common.0x6FD520F0 (#11090)
-BOOL __stdcall DATATBLS_FreeCollisionFieldTable();
-//D2Common.0x6FD52140 (#11091)
-void __stdcall DATATBLS_GetCollisionFieldCenter(int* a1, int* a2);
-//D2Common.0x6FD52160 (#11092)
-int __stdcall DATATBLS_GetCollisionFieldWidth();
-//D2Common.0x6FD52170 (#11093)
-int __stdcall DATATBLS_GetCollisionFieldHeight();
-//D2Common.0x6FD52180 (#11094)
-D2FieldStrc* __stdcall DATATBLS_AllocField();
-//D2Common.0x6FD521D0 (#11095)
-void __stdcall DATATBLS_FreeField(D2FieldStrc* pField);
-//D2Common.0x6FD52210 (#11096)
-void __stdcall DATATBLS_SetFieldCoordinates(D2FieldStrc* pField, int nX, int nY);
-//D2Common.0x6FD52250 (#11097)
-int __stdcall D2Common_11097(D2FieldStrc* pField, int a2, int a3);
-//D2Common.0x6FD522A0 (#11098)
-int __stdcall D2Common_11098(D2FieldStrc* pField, int* pX, int* pY);
-//D2Common.0x6FD52360 (#11099)
-BOOL __stdcall D2Common_11099(D2FieldStrc* pField, D2RoomStrc* pRoom, int nX, int nY, uint16_t fMask);
-
-
-
-//D2Common.0x6FD523E0
-void __fastcall DATATBLS_CubeMainInputLinker(char* pSrc, void* pRecord, int nOffset, int nPosition, int nTxtRow, int nTxtColumn);
-//D2Common.0x6FD52410
-BOOL __fastcall DATATBLS_CubeMainInputParser(D2CubeInputItem* pCubeInput, char* szInput, int a3, int a4);
-//D2Common.0x6FD528D0
-void __fastcall DATATBLS_CubeMainOutputLinker(char* pSrc, void* pRecord, int nOffset, int nPosition, int nTxtRow, int nTxtColumn);
-//D2Common.0x6FD52910
-BOOL __fastcall DATATBLS_CubeMainOutputParser(D2CubeOutputItem* pCubeOutputParam, char* szOutput, int nTxtRow, int nItemId);
-//D2Common.0x6FD52FC0
-void __fastcall DATATBLS_CubeMainParamLinker(char* pSrc, void* pRecord, int nOffset, int nPosition, int nTxtRow, int nTxtColumn);
-//D2Common.0x6FD53030
-void __fastcall DATATBLS_LoadCubeMainTxt(void* pMemPool);
-//D2Common.0x6FD54250
-void __fastcall DATATBLS_UnloadCubeMainTxt();
-//D2Common.0x6FD54260 (#11232)
-D2CubeMainTxt* __stdcall DATATBLS_GetCubemainTxtRecord(int nIndex);
-//D2Common.0x6FD542C0 (#11233)
-int __fastcall DATATBLS_GetCubemainTxtRecordCount();
-
-
-
-//D2Common.0x6FD542D0
-void __fastcall DATATBLS_LoadInventoryTxt(void* pMemPool);
-//D2Common.0x6FD54F10
-void __fastcall DATATBLS_UnloadInventoryTxt();
-//D2Common.0x6FD54F20 (#10635)
-void __stdcall DATATBLS_GetInventoryRect(int nInventoryTxtId, int bHigherRes, D2InvRectStrc* pInvRect);
-//D2Common.0x6FD54FB0 (#10636)
-void __stdcall DATATBLS_GetInventoryGridInfo(int nInventoryTxtId, int bHigherRes, D2InventoryGridInfoStrc* pInventoryGridInfo);
-//D2Common.0x6FD55030 (#10637)
-void __stdcall DATATBLS_GetInventoryComponentGrid(int nInventoryTxtId, int bHigherRes, D2InvCompGridStrc* pInvCompGrid, int nComponent);
-//D2Common.0x6FD550E0
-int __fastcall DATATBLS_MapItemsTxtKeywordToNumber(char* szKey);
-//D2Common.0x6FD55140
-int __fastcall DATATBLS_Return2();
-//D2Common.0x6FD55150
-int __fastcall sub_6FD55150(char* szText, int* a2, int a3, int nKeywordNumber);
-//D2Common.0x6FD551E0
-void __fastcall DATATBLS_ItemCalcLinker(char* pSrc, void* pRecord, int nOffset, int nPosition, int nTxtRow, int nTxtColumn);
-//D2Common.0x6FD55280
-void __fastcall DATATBLS_LoadItemsTxt(void* pMemPool);
-//D2Common.0x6FD575D0
-void __fastcall DATATBLS_UnloadItemsTxt();
-//D2Common.0x6FD57620 (#10599)
-D2ItemDataTbl* __stdcall DATATBLS_GetItemDataTables();
-//D2Common.0x6FD57630 (#10597)
-int __stdcall DATATBLS_MapOldItemIndexToCurrent(int nItemId);
-//D2Common.0x6FD57680 (#10600)
-D2ItemsTxt* __stdcall DATATBLS_GetItemsTxtRecord(int nItemId);
-//D2Common.0x6FD576D0 (#10601)
-D2ItemsTxt* __stdcall DATATBLS_GetItemRecordFromItemCode(uint32_t dwCode, int* pItemId);
-//D2Common.0x6FD57720 (#10602)
-int __stdcall DATATBLS_GetItemIdFromItemCode(uint32_t dwCode);
-//D2Common.0x6FD57740
-void __fastcall DATATBLS_ItemParamLinker(char* pSrc, void* pRecord, int nOffset, int nPosition, int nTxtRow, int nTxtColumn);
-//D2Common.0x6FD57820
-void __fastcall DATATBLS_LoadMagicSuffix_Prefix_AutomagicTxt(void* pMemPool);
-//D2Common.0x6FD58080
-void __fastcall DATATBLS_UnloadMagicSuffix_Prefix_AutomagicTxt();
-//D2Common.0x6FD580B0 (#10603)
-D2MagicAffixDataTbl* __stdcall DATATBLS_GetMagicAffixDataTables();
-//D2Common.0x6FD580C0 (#10604)
-D2MagicAffixTxt* __stdcall DATATBLS_GetMagicAffixTxtRecord(int nIndex);
-//D2Common.0x6FD58110
-void __fastcall DATATBLS_LoadRareSuffix_PrefixTxt(void* pMemPool);
-//D2Common.0x6FD58450
-void __fastcall DATATBLS_UnloadRareSuffix_PrefixTxt();
-//D2Common.0x6FD58480 (#10605)
-D2RareAffixDataTbl* __fastcall DATATBLS_GetRareAffixDataTables();
-//D2Common.0x6FD58490 (#10606)
-D2RareAffixTxt* __stdcall DATATBLS_GetRareAffixTxtRecord(int nId);
-//D2Common.0x6FD584E0
-void __fastcall DATATBLS_LoadUniqueItemsTxt(void* pMemPool);
-//D2Common.0x6FD59110
-void __fastcall DATATBLS_UnloadUniqueItemsTxt();
-//D2Common.0x6FD59140
-void __fastcall DATATBLS_LoadSets_SetItemsTxt(void* pMemPool);
-//D2Common.0x6FD5AE00
-void __fastcall DATATBLS_UnloadSets_SetItemsTxt();
-//D2Common.0x6FD5AE40
-void __fastcall DATATBLS_LoadQualityItemsTxt(void* pMemPool);
-//D2Common.0x6FD5B250
-void __fastcall DATATBLS_UnloadQualityItemsTxt();
-//D2Common.0x6FD5B260 (#10611)
-D2QualityItemDataTbl* __fastcall DATATBLS_GetQualityItemDataTables();
-//D2Common.0x6FD5B270 (#10612)
-D2QualityItemsTxt* __stdcall DATATBLS_GetQualityItemsTxtRecord(int nIndex);
-//D2Common.0x6FD5B2F0
-void __fastcall DATATBLS_LoadGemsTxt(void* pMemPool);
-//D2Common.0x6FD5BAE0
-void __fastcall DATATBLS_UnloadGemsTxt();
-//D2Common.0x6FD5BAF0 (#10615)
-D2GemDataTbl* __fastcall DATATBLS_GetGemDataTables();
-//D2Common.0x6FD5BB00 (#10616)
-D2GemsTxt* __stdcall DATATBLS_GetGemsTxtRecord(int nGemId);
-//D2Common.0x6FD5BB70
-void __fastcall DATATBLS_LoadBooksTxt(void* pMemPool);
-//D2Common.0x6FD5BD10
-void __fastcall DATATBLS_UnloadBooksTxt();
-//D2Common.0x6FD5BD20 (#10617)
-D2BookDataTbl* __fastcall DATATBLS_GetBookDataTables();
-//D2Common.0x6FD5BD30 (#10618)
-D2BooksTxt* __stdcall DATATBLS_GetBooksTxtRecord(int nBookId);
-//D2Common.0x6FD5BDA0
-void __fastcall DATATBLS_LoadLowQualityItemsTxt(void* pMemPool);
-//D2Common.0x6FD5BE40
-void __fastcall DATATBLS_UnloadLowQualityItemsTxt();
-//D2Common.0x6FD5BE50 (#10613)
-D2LowQualityItemDataTbl* __fastcall DATATBLS_GetLowQualityItemDataTables();
-//D2Common.0x6FD5BE60 (#10614)
-D2LowQualityItemsTxt* __stdcall DATATBLS_GetLowQualityItemsTxtRecord(int nId);
-//D2Common.0x6FD5BED0
-void __fastcall DATATBLS_LoadItemRatioTxt(void* pMemPool);
-//D2Common.0x6FD5C200
-void __fastcall DATATBLS_UnloadItemRatioTxt();
-//D2Common.0x6FD5C210 (#10622)
-D2ItemRatioDataTbl* __fastcall DATATBLS_GetItemRatioDataTables();
-//D2Common.0x6FD5C220 (#10623)
-D2ItemRatioTxt* __stdcall DATATBLS_GetItemRatioTxtRecord(int nItemId, uint8_t nDifficulty, uint16_t wVersion);
-//D2Common.0x6FD5C2F0
-int __cdecl DATATBLS_CompareItemStatCostDescs(const void* pRecord1, const void* pRecord2);
-//D2Common.0x6FD5C320
-void __fastcall DATATBLS_LoadItemStatCostTxt(void* pMemPool);
-//D2Common.0x6FD5D070
-void __fastcall DATATBLS_UnloadItemStatCostTxt();
-//D2Common.0x6FD5D0D0
-void __fastcall DATATBLS_LoadPropertiesTxt(void* pMemPool);
-//D2Common.0x6FD5D5E0
-void __fastcall DATATBLS_UnloadPropertiesTxt();
-//D2Common.0x6FD5D600
-void __fastcall DATATBLS_LoadGambleTxt(void* pMemPool);
-//D2Common.0x6FD5D790
-int __cdecl DATATBLS_CompareGambleTxtRecords(const void* pRecord1, const void* pRecord2);
-//D2Common.0x6FD5D7B0
-void __fastcall DATATBLS_UnloadGambleTxt();
-//D2Common.0x6FD5D7F0 (#10671)
-D2GambleDataTbl* __fastcall DATATBLS_GetGambleDataTables();
-//D2Common.0x6FD5D800
-BOOL __fastcall DATATBLS_CheckNestedItemTypes(int nItemType1, int nItemType2);
-//D2Common.0x6FD5D8C0
-void __fastcall DATATBLS_LoadItemTypesTxt(void* pMemPool);
-//D2Common.0x6FD5DFE0
-void __fastcall DATATBLS_UnloadItemTypesTxt();
-//D2Common.0x6FD5E020
-void __fastcall DATATBLS_LoadRunesTxt(void* pMemPool);
-//D2Common.0x6FD5E9C0
-void __fastcall DATATBLS_UnloadRunesTxt();
-//D2Common.0x6FD5E9E0 (#10619)
-D2RuneDataTbl* __fastcall DATATBLS_GetRuneDataTables();
-//D2Common.0x6FD5E9F0 (#10621)
-void __stdcall DATATBLS_AddOrChangeRunesTxtRecord(int nRecordId, D2RunesTxt* pRecord);
-//D2Common.0x6FD5EAA0 (#10620)
-D2RunesTxt* __stdcall DATATBLS_GetRunesTxtRecord(int nRunewordId);
-
-
-
-//D2Common.0x6FD5EAE0
-void __fastcall DATATBLS_LoadLevelsTxt(void* pMemPool);
-//D2Common.0x6FD603C0 (#10631)
-D2LevelsTxt* __stdcall DATATBLS_GetLevelsTxtRecord(int nLevelId);
-//D2Common.0x6FD603F0 (#10632)
-uint8_t __stdcall DATATBLS_GetRainFromLevelsTxt(int nLevelId);
-//D2Common.0x6FD60430 (#10634)
-uint8_t __stdcall DATATBLS_GetNoPerFromLevelsTxt(int nLevelId);
-//D2Common.0x6FD60470 (#10633)
-uint8_t __stdcall DATATBLS_GetMudFromLevelsTxt(int nLevelId);
-//D2Common.0x6FD604B0
-void __fastcall DATATBLS_UnloadLevelsTxt();
-//D2Common.0x6FD604F0 (#11247)
-int __stdcall DATATBLS_GetMonsterLevelInArea(int nLevelId, uint8_t nDifficulty, BOOL bExpansion);
-//D2Common.0x6FD60560
-int* __fastcall DATATBLS_GetPortalLevels(int* pnPortalLevels);
-//D2Common.0x6FD60570
-void __fastcall DATATBLS_LoadLevelDefsBin(void* pMemPool);
-//D2Common.0x6FD60D60
-void __fastcall DATATBLS_UnloadLevelDefsBin();
-//D2Common.0x6FD60D90 (#10010)
-D2LevelDefBin* __fastcall DATATBLS_GetLevelDefRecord(int nLevelId);
-//D2Common.0x6FD60DC0
-void __fastcall DATATBLS_LoadLevelTypesTxt(void* pMemPool);
-//D2Common.0x6FD61450
-void __fastcall DATATBLS_UnloadLevelTypesTxt();
-//D2Common.0x6FD61460 (#10023)
-D2LvlTypesTxt* __stdcall DATATBLS_GetLevelTypesTxtRecord(int nLevelType);
-//D2Common.0x6FD614A0 (#11226)
-bool __stdcall DATATBLS_CheckActInLevelTypesTxt(int nLevelType, uint8_t nAct);
-//D2Common.0x6FD61500 (#11227)
-void __stdcall DATATBLS_GetFileNameFromLevelTypeAndFileId(int nLevelType, int nFile, char* szFile);
-//D2Common.0x6FD61570
-void __fastcall DATATBLS_LoadLvlPrestTxt(void* pMemPool, int a2);
-//D2Common.0x6FD61AD0
-void __fastcall DATATBLS_UnloadLvlPrestTxt();
-//D2Common.0x6FD61B50 (#10024)
-D2LvlPrestTxt* __stdcall DATATBLS_GetLvlPrestTxtRecord(int nId);
-//D2Common.0x6FD61B80
-D2LvlPrestTxt* __fastcall DATATBLS_GetLvlPrestTxtRecordFromLevelId(int nLevelId);
-//D2Common.0x6FD61BC0
-void __fastcall DATATBLS_LoadLvlWarpTxt(void* pMemPool);
-//D2Common.0x6FD61DD0
-void __fastcall DATATBLS_UnloadLvlWarpTxt();
-//D2Common.0x6FD61DE0
-D2LvlWarpTxt* __fastcall DATATBLS_GetLvlWarpTxtRecordFromLevelIdAndDirection(int nLevelId, char szDirection);
-//D2Common.0x6FD61E60
-void __fastcall DATATBLS_LoadLvlMazeTxt(void* pMemPool);
-//D2Common.0x6FD61F90
-void __fastcall DATATBLS_UnloadLvlMazeTxt();
-//D2Common.0x6FD61FA0
-D2LvlMazeTxt* __fastcall DATATBLS_GetLvlMazeTxtRecordFromLevelId(int nLevelId);
-//D2Common.0x6FD62020
-void __fastcall DATATBLS_LoadLvlSubTxt(void* pMemPool, int a2, int a3);
-//D2Common.0x6FD62600
-void __fastcall DATATBLS_UnloadLvlSubTxt();
-//D2Common.0x6FD626F0
-D2LvlSubTxt* __fastcall DATATBLS_GetLvlSubTxtRecord(int nSubType);
-//D2Common.0x6FD62780
-void __fastcall DATATBLS_AllocGlobalTileLibraryHash();
-//D2Common.0x6FD628A0
-void __fastcall DATATBLS_FreeGlobalTileLibraryHash();
-//D2Common.0x6FD628C0
-void __fastcall DATATBLS_LoadAutomapTxt(void* pMemPool);
-//D2Common.0x6FD62D30 (#10011)
-int __fastcall DATATBLS_GetAutomapCellId(uint32_t dwAutomapLevelType, uint32_t dwAutomapTileType, int nStyle, int nSequence);
-//D2Common.0x6FD62E70
-void __fastcall DATATBLS_FreeAutomap();
-
-
-
-//D2Common.0x6FD62EA0
-int __fastcall DATATBLS_MapMissilesTxtKeywordToNumber(char* szKey);
-//D2Common.0x6FD62F20
-int __fastcall sub_6FD62F20(char* szText, int* a2, int a3, int nKeywordNumber);
-//D2Common.0x6FD630F0
-void __fastcall DATATBLS_MissileCalcLinker(char* pSrc, void* pRecord, int nOffset, int nPosition, int nTxtRow, int nTxtColumn);
-//D2Common.0x6FD63180
-void __fastcall DATATBLS_LoadMissilesTxt(void* pMemPool);
-//D2Common.0x6FD64B80
-void __fastcall DATATBLS_UnloadMissilesTxt();
-//D2Common.0x6FD64BE0 (#10590)
-int __stdcall DATATBLS_GetMissileVelocityFromMissilesTxt(int nMissileId, int nLevel);
-//Inlined at various places
-D2MissilesTxt* __fastcall DATATBLS_GetMissilesTxtRecord(int nMissileId);
-
-
-//Inlined in some functions
-uint32_t __fastcall DATATBLS_StringToCode(char* szText);
-//D2Common.0x6FD64C40 (#10592)
-int __stdcall DATATBLS_GetMonsterChainInfo(int nMonsterId, int* pMaxChainId, int* pChainId);
-//D2Common.0x6FD64CD0
-void __fastcall DATATBLS_MonStatsSkillModeLinker(char* pSrc, void* pRecord, int nOffset, int nPosition, int nTxtRow, int nTxtColumn);
-//D2Common.0x6FD64DF0
-void __fastcall DATATBLS_LoadMonStatsTxt(void* pMemPool);
-//D2Common.0x6FD67BD0 (#11261)
-BOOL __stdcall DATATBLS_CalculateMonsterStatsByLevel(int nMonsterId, int nGameType, int nDifficulty, int nLevel, short nFlags, D2MonStatsInitStrc* pMonStatsInit);
-//D2Common.0x6FD689B0 (#10642)
-void __stdcall DATATBLS_SetVelocityInMonStatsTxtRecord(int nMonsterId, short nVelocity);
-//D2Common.0x6FD68A00
-void __stdcall DATATBLS_ResetGlobalDefaultUnicodeString();
-//D2Common.0x6FD68A10 (#10651)
-wchar_t* __fastcall DATATBLS_RollRandomUniqueTitleString(D2UnitStrc* pUnit);
-//D2Common.0x6FD68A80 (#10652)
-wchar_t* __fastcall DATATBLS_RollRandomUniquePrefixString(D2UnitStrc* pUnit);
-//D2Common.0x6FD68AF0 (#10653)
-wchar_t* __fastcall DATATBLS_RollRandomUniqueSuffixString(D2UnitStrc* pUnit);
-//D2Common.0x6FD68B60 (#10654)
-wchar_t* __fastcall DATATBLS_RollRandomUniqueAppellationString(D2UnitStrc* pUnit);
-//D2Common.0x6FD68BD0
-void __fastcall DATATBLS_ReallocTCExInfo(D2TCExShortStrc* pTCExShort, int nNewRecordCount);
-//D2Common.0x6FD68C40
-void __fastcall DATATBLS_UpdateTreasureClassProbabilities(D2TCExShortStrc* pTCExShort, D2TCExInfoStrc* pTCExInfo, int nProbability, BOOL bExpansion);
-//D2Common.0x6FD68D40
-int __fastcall DATATBLS_CalculatePercentage(signed int nValue1, signed int nValue2, signed int nDivisor);
-//D2Common.0x6FD68DC0 (#10658)
-D2TCExShortStrc* __stdcall DATATBLS_GetTreasureClassExRecordFromName(char* szText);
-//D2Common.0x6FD68DF0 (#10659)
-D2TCExShortStrc* __stdcall DATATBLS_GetTreasureClassExRecordFromIdAndLevel(uint16_t wTCId, int nLvl);
-//D2Common.0x6FD68E50 (#10660)
-D2TCExShortStrc* __stdcall DATATBLS_GetTreasureClassExRecordFromActAndDifficulty(int nDifficulty, int nAct, int nIndex);
-//D2Common.0x6FD68EC0
-void __fastcall DATATBLS_LoadTreasureClassExTxt(void* pMemPool);
-//D2Common.0x6FD69B70 (#10656)
-int __stdcall DATATBLS_ShouldNeverCallInExpansion();
-//D2Common.0x6FD69B90
-void __fastcall DATATBLS_LoadMonItemPercentTxt(void* pMemPool);
-//D2Common.0x6FD69C40
-void __fastcall DATATBLS_UnloadMonItemPercentTxt();
-//D2Common.0x6FD69C50 (#10662)
-D2MonItemPercentDataTbl* __fastcall DATATBLS_GetMonItemPercentDataTables();
-//D2Common.0x6FD69C60 (#10663)
-D2MonItemPercentTxt* __stdcall DATATBLS_GetMonItemPercentTxtRecord(int nId);
-//D2Common.0x6FD69CD0
-void __fastcall DATATBLS_LoadMonUModTxt(void* pMemPool);
-//D2Common.0x6FD69FA0
-void __fastcall DATATBLS_LoadSuperUniquesTxt(void* pMemPool);
-//D2Common.0x6FD6A440 (#10668)
-D2SuperUniquesTxt* __stdcall DATATBLS_GetSuperUniquesTxtRecord(int nSuperUniqueId);
-//D2Common.0x6FD6A470 (#11257)
-int __fastcall DATATBLS_GetSuperUniquesTxtRecordCount();
-//D2Common.0x6FD6A480
-void __fastcall DATATBLS_LoadHirelingTxt(void* pMemPool);
-//D2Common.0x6FD6B1A0
-D2ItemTypesTxt* __fastcall DATATBLS_GetItemTypesTxtRecord(int nItemType);
-//D2Common.0x6FD6B1D0 (#10583)
-D2HirelingTxt* __stdcall DATATBLS_GetHirelingTxtRecordFromIdAndLevel(BOOL bExpansion, int nId, int nLevel);
-//D2Common.0x6FD6B270 (#10585)
-D2HirelingTxt* __stdcall DATATBLS_GetNextHirelingTxtRecordFromNameId(BOOL bExpansion, uint16_t nNameId, D2HirelingTxt* pOldRecord);
-//D2Common.0x6FD6B310 (#10586)
-D2HirelingTxt* __stdcall DATATBLS_GetNextHirelingTxtRecordFromClassId(BOOL bExpansion, int nClass, D2HirelingTxt* pOldRecord);
-//D2Common.0x6FD6B3A0 (#10587)
-D2HirelingTxt* __stdcall DATATBLS_GetNextHirelingTxtRecordFromVendorIdAndDifficulty(BOOL bExpansion, int nVendorId, int nDifficulty, D2HirelingTxt* pOldRecord);
-//D2Common.0x6FD6B440 (#10584)
-D2HirelingTxt* __stdcall DATATBLS_GetNextHirelingTxtRecordFromActAndDifficulty(BOOL bExpansion, int nAct, int nDifficulty, D2HirelingTxt* pOldRecord);
-//D2Common.0x6FD6B4F0
-void __fastcall DATATBLS_LoadNpcTxt(void* pMemPool);
-//D2Common.0x6FD6B820 (#10588)
-D2NpcTxt* __stdcall DATATBLS_GetNpcTxtRecord(uint32_t dwNpcId);
-//D2Common.0x6FD6B850
-void __fastcall DATATBLS_LoadMonSoundsTxt(void* pMemPool);
-//D2Common.0x6FD6BF50 (#11252)
-D2MonSoundsTxt* __stdcall DATATBLS_GetMonSoundsTxtRecordFromSoundId(int nSoundId);
-//D2Common.0x6FD6BF80 (#11250)
-D2MonSoundsTxt* __stdcall DATATBLS_GetMonSoundsTxtRecordFromMonsterId(int nMonsterId);
-//D2Common.0x6FD6BFE0
-void __fastcall DATATBLS_MonStats2CompositLinker(char* pSrc, void* pRecord, int nOffset, int nPosition, int nTxtRow, int nTxtColumn);
-//D2Common.0x6FD6C150
-void __fastcall DATATBLS_LoadMonStats2Txt(void* pMemPool);
-//D2Common.0x6FD6D660
-int __fastcall DATATBLS_CheckNestedMonsterTypes(int nMonType1, int nMonType2);
-//D2Common.0x6FD6D730
-void __fastcall DATATBLS_LoadMonTypeTxt(void* pMemPool);
-//D2Common.0x6FD6D910
-void __fastcall DATATBLS_UnloadMonTypeTxt();
-//D2Common.0x6FD6D960
-void __fastcall DATATBLS_LoadMonPropTxt(void* pMemPool);
-//D2Common.0x6FD6E8E0
-void __fastcall DATATBLS_LoadMonLvlTxt(void* pMemPool);
-//D2Common.0x6FD6EDE0
-void __fastcall DATATBLS_MonPresetPlaceLinker(char* pSrc, void* pRecord, int nOffset, int nPosition, int nTxtRow, int nTxtColumn);
-//D2Common.0x6FD6EE60
-void __fastcall DATATBLS_LoadMonPresetTxt(void* pMemPool);
-//D2Common.0x6FD6EF30 (#11256)
-D2MonPresetTxt* __stdcall DATATBLS_GetMonPresetTxtActSection(int nAct, int* pRecordCount);
-//D2Common.0x6FD6EF70 (#11258)
-int __stdcall DATATBLS_MapSuperUniqueId(int nType, int nSuperUnique);
-//D2Common.0x6FD6EFA0
-void __fastcall DATATBLS_LoadSoundsTxt(void* pMemPool);
-//D2Common.0x6FD6F020
-void __fastcall DATATBLS_UnloadSoundsTxt();
-//D2Common.0x6FD6F050
-void __fastcall DATATBLS_LoadMonSeqTxt(void* pMemPool);
-//D2Common.0x6FD6F200 (#11262)
-D2SeqRecordStrc* __stdcall DATATBLS_GetMonSeqTableRecord(int nSequence);
-//D2Common.0x6FD6F230
-void __fastcall DATATBLS_LoadMonEquipTxt(void* pMemPool);
-//D2Common.0x6FD6F580
-void __fastcall DATATBLS_LoadSomeMonsterTxts(void* pMemPool);
-//D2Common.0x6FD6F810
-void __fastcall DATATBLS_CreateItemTypeTreasureClasses();
-//D2Common.0x6FD6FBB0
-void __fastcall DATATBLS_UnloadSomeMonsterTxts();
-//Inlined at various places
-D2MonStatsTxt* __fastcall DATATBLS_GetMonStatsTxtRecord(int nMonsterId);
-
-
-
-//D2Common.0x6FD6FDF0
-void __fastcall DATATBLS_LoadObjectsTxt(void* pMemPool);
-//D2Common.0x6FD718F0 (#10626)
-D2ObjectsTxt* __stdcall DATATBLS_GetObjectsTxtRecord(int nObjectId);
-//D2Common.0x6FD71960
-void __fastcall DATATBLS_UnloadObjectsTxt();
-//D2Common.0x6FD71980
-void __fastcall DATATBLS_UnloadObjGroupTxt();
-//D2Common.0x6FD719A0
-void __fastcall DATATBLS_LoadObjGroupTxt(void* pMemPool);
-//D2Common.0x6FD71E00 (#10627)
-D2ObjGroupTxt* __stdcall DATATBLS_GetObjGroupTxtRecord(int nId);
-//D2Common.0x6FD71E30
-void __fastcall DATATBLS_LoadShrinesTxt(void* pMemPool);
-//D2Common.0x6FD72000 (#10624)
-D2ShrinesTxt* __stdcall DATATBLS_GetShrinesTxtRecord(int nShrineId);
-//D2Common.0x6FD72070 (#10625)
-int __stdcall DATATBLS_GetShrinesTxtRecordCount();
-//D2Common.0x6FD72080
-void __fastcall DATATBLS_UnloadShrinesTxt();
-
-
-
-//D2Common.0x6FD720A0
-void __fastcall DATATBLS_LoadOverlayTxt(void* pMemPool);
-//D2Common.0x6FD72500
-void __fastcall DATATBLS_UnloadOverlayTxt();
-//D2Common.0x6FD72530 (#10674)
-uint32_t __stdcall DATATBLS_GetFramesFromOverlayTxt(int nOverlayId);
-//D2Common.0x6FD72570 (#10675)
-int __stdcall DATABLS_GetPreDrawFromOverlayTxt(int nOverlayId);
-//D2Common.0x6FD725B0 (#10676)
-uint32_t __stdcall DATATBLS_GetOffsetXFromOverlayTxt(int nOverlayId);
-//D2Common.0x6FD725F0 (#10677)
-uint32_t __stdcall DATATBLS_GetOffsetYFromOverlayTxt(int nOverlayId);
-//D2Common.0x6FD72630 (#10678)
-int __stdcall DATATBLS_GetTransFromOverlayTxt(int nOverlayId);
-//D2Common.0x6FD72670 (#10679)
-int __stdcall DATATBLS_GetRadiusAndColorFromOverlayTxt(int nOverlayId, int* pInitRadius, int* pRadius, uint8_t* pRed, uint8_t* pGreen, uint8_t* pBlue);
-//D2Common.0x6FD72720 (#10680)
-uint32_t __stdcall DATATBLS_Get1OfNFromOverlayTxt(int nOverlayId);
-//D2Common.0x6FD72760 (#10681)
-int __stdcall DATATBLS_GetDirFromOverlayTxt(int nOverlayId);
-//Inlined at various places
-D2OverlayTxt* __fastcall DATATBLS_GetOverlayTxtRecord(int nOverlay);
-
-
-
-//D2Common.0x6FD727A0 (#10682)
-D2MonSeqTxt* __stdcall DATATBLS_GetMonSeqTxtRecordFromUnit(D2UnitStrc* pUnit);
-//D2Common.0x6FD727C0
-D2SeqRecordStrc* __fastcall DATATBLS_GetSeqRecordFromUnit(D2UnitStrc* pUnit);
-//D2Common.0x6FD728A0 (#10683)
-int __stdcall D2Common_10683(D2UnitStrc* pUnit);
-//D2Common.0x6FD728C0 (#10684)
-int __stdcall D2Common_10684(D2UnitStrc* pUnit);
-//D2Common.0x6FD728E0 (#10685)
-void __stdcall D2Common_10685(D2MonSeqTxt* pMonSeqTxt, int a2, int a3, unsigned int* pMode, unsigned int* pFrame, int* pDirection, int* pEvent);
-//D2Common.0x6FD72990 (#10686)
-void __stdcall DATATBLS_GetSequenceEvent(D2MonSeqTxt* pMonSeqTxt, int a2, int* pEvent);
-//D2Common.0x6FD729C0
-void __fastcall DATATBLS_LoadPlrType_ModeTxt(void* pMemPool);
-//D2Common.0x6FD72B30
-void __fastcall DATATBLS_LoadMonModeTxt(void* pMemPool);
-//D2Common.0x6FD72E50
-void __fastcall DATATBLS_LoadObjType_ModeTxt(void* pMemPool);
-//D2Common.0x6FD72FC0
-void __fastcall DATATBLS_LoadCompositTxt(void* pMemPool);
-//D2Common.0x6FD73040
-void __fastcall DATATBLS_LoadArmTypeTxt(void* pMemPool);
-//D2Common.0x6FD730C0
-void __fastcall DATATBLS_UnloadPlrMode_Type_MonMode_ObjMode_Type_Composit_ArmtypeTxt();
-//D2Common.0x6FD73150 (#10643)
-D2PlrModeDataTbl* __fastcall DATATBLS_GetPlrMode_TypeDataTables();
-//D2Common.0x6FD73160 (#10644)
-D2MonModeDataTbl* __fastcall DATATBLS_GetMonModeDataTables();
-//D2Common.0x6FD73170 (#10645)
-D2ObjModeDataTbl* __fastcall DATATBLS_GetObjMode_TypeDataTables();
-//D2Common.0x6FD73180 (#10646)
-D2PlrModeTypeTxt* __stdcall DATATBLS_GetPlrModeTypeTxtRecord(int nIndex, int bGetMode);
-//D2Common.0x6FD73230 (#10647)
-D2MonModeTxt* __stdcall DATATBLS_GetMonModeTxtRecord(int nIndex, int bGetMode);
-//D2Common.0x6FD732B0 (#10648)
-D2ObjModeTypeTxt* __stdcall DATATBLS_GetObjModeTypeTxtRecord(int nIndex, int bGetMode);
-//D2Common.0x6FD73330 (#10649)
-D2CompositTxt* __stdcall DATATBLS_GetCompositTxtRecord(int nComposit);
-//D2Common.0x6FD73370 (#10650)
-D2ArmTypeTxt* __stdcall DATATBLS_GetArmTypeTxtRecord(int nId);
-//D2Common.0x6FD733B0 (#10667)
-int __stdcall DATATBLS_UNUSED_Return0(int a1, int a2);
