+++ source/D2Common/include/D2ItemMods.h
@@ -2,6 +2,23 @@
 
 #include "CommonDefinitions.h"
 #include <Units/Missile.h>
+#include <DataTbls/ItemsTbls.h>
+
+//struct D2PropertySetFuncStrc
+//{
+//	int32_t nState;
+//	int32_t fStatList;
+//};
+
+typedef BOOL(__fastcall* PROPERTYASSIGN)(int32_t, D2UnitStrc*, D2UnitStrc*, D2PropertyStrc*, int32_t, int32_t, int32_t, int32_t, D2UnitStrc*);
+
+typedef int32_t(__fastcall* PROPERTYASSIGNFN)(int32_t, D2UnitStrc*, D2UnitStrc*, const D2PropertyStrc*, int32_t, int16_t, int32_t, int32_t, int32_t, int32_t, D2UnitStrc*);
+
+struct D2PropertyAssignStrc
+{
+	PROPERTYASSIGN pfAssign;				//0x00
+	int32_t nStatId;							//0x04
+};
 
 //D2Common.0x6FD92640 (#10844)
 void __stdcall D2Common_10844_ITEMMODS_First(int nDataBits, int* pLayer, int* pValue);
