+++ source/D2Common/include/D2Items.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "CommonDefinitions.h"
-
+#include <DataTbls/ItemsTbls.h>
 //TODO: Revise
