+++ source/D2Common/include/DataTbls/DataTbls.h
@@ -1,0 +1,3 @@
+#pragma once 
+
+#include <D2BasicTypes.h>
