+++ source/D2Common/include/DataTbls/InvTbls.h
@@ -1,0 +1,46 @@
+#pragma once 
+
+#include <D2BasicTypes.h>
+#include <D2Inventory.h>
+
+
+
+struct D2InvCompGridStrc
+{
+	D2InvRectStrc rect;						//0x00
+	uint8_t nWidth;							//0x10
+	uint8_t nHeight;						//0x11
+	uint16_t pad0x12;						//0x12
+};
+
+struct D2InventoryGridInfoStrc
+{
+	uint8_t nGridX;							//0x00
+	uint8_t nGridY;							//0x01
+	uint16_t pad0x02;						//0x02
+	int32_t nGridLeft;						//0x04
+	int32_t nGridRight;						//0x08
+	int32_t nGridTop;						//0x0C
+	int32_t nGridBottom;					//0x10
+	uint8_t nGridBoxWidth;					//0x14
+	uint8_t nGridBoxHeight;					//0x15
+	uint16_t pad0x16;						//0x16
+};
+
+struct D2InventoryTxt
+{
+	D2InvRectStrc pRect;					//0x00
+	D2InventoryGridInfoStrc pGridInfo;		//0x10
+	D2InvCompGridStrc pComponents[10];		//0x28
+};
+
+//D2Common.0x6FD542D0
+void __fastcall DATATBLS_LoadInventoryTxt(void* pMemPool);
+//D2Common.0x6FD54F10
+void __fastcall DATATBLS_UnloadInventoryTxt();
+//D2Common.0x6FD54F20 (#10635)
+void __stdcall DATATBLS_GetInventoryRect(int nInventoryTxtId, int bHigherRes, D2InvRectStrc* pInvRect);
+//D2Common.0x6FD54FB0 (#10636)
+void __stdcall DATATBLS_GetInventoryGridInfo(int nInventoryTxtId, int bHigherRes, D2InventoryGridInfoStrc* pInventoryGridInfo);
+//D2Common.0x6FD55030 (#10637)
+void __stdcall DATATBLS_GetInventoryComponentGrid(int nInventoryTxtId, int bHigherRes, D2InvCompGridStrc* pInvCompGrid, int nComponent);
