+++ source/D2Common/include/DataTbls/TransformTbls.h
@@ -1,0 +1,6 @@
+#pragma once 
+
+#include <D2BasicTypes.h>
+
+//D2Common.0x6FD733B0 (#10667)
+int __stdcall DATATBLS_UNUSED_Return0(int a1, int a2);
