+++ source/D2Common/include/Drlg/D2DrlgDrlgLogic.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "CommonDefinitions.h"
+#include "D2DrlgDrlgGrid.h"
 
 struct D2UnkDrlgLogicStrc
 {
@@ -12,6 +13,15 @@ struct D2UnkDrlgLogicStrc
 	D2DrlgGridStrc* field_14;				//0x14
 	int32_t field_18;						//0x18
 	int32_t nFlags;							//0x1C
+};
+
+struct D2DrlgCoordListStrc
+{
+	int32_t dwFlags;							//0x00
+	int32_t nLists;								//0x04
+	D2DrlgGridStrc pIndexX;					//0x08
+	D2DrlgGridStrc pIndexY;					//0x1C
+	D2RoomCoordListStrc* pCoordList;		//0x30
 };
 
 //D2Common.0x6FD76420
