+++ source/D2Common/include/Drlg/D2DrlgOutdoors.h
@@ -1,6 +1,38 @@
 #pragma once
 
 #include "CommonDefinitions.h"
+#include "D2DrlgDrlgGrid.h"
+
+struct D2DrlgOutdoorGridStrc
+{
+	int32_t dwFlags;						//0x00
+	D2DrlgGridStrc* pSectors;				//0x04
+	int32_t nWidth;							//0x08
+	int32_t nHeight;						//0x0C
+	BOOL bInit;								//0x10
+};
+
+struct D2DrlgOutdoorInfoStrc
+{
+	uint32_t dwFlags;							//0x00
+	D2DrlgGridStrc pGrid[4];				//0x04
+	union
+	{
+		struct
+		{
+			int32_t nWidth;								//0x54
+			int32_t nHeight;							//0x58
+			int32_t nGridWidth;							//0x5C
+			int32_t nGridHeight;						//0x60
+		};
+		D2DrlgCoordStrc pCoord;
+	};
+	D2DrlgVertexStrc* pVertex;				//0x64
+	D2DrlgVertexStrc* unk0x68[6];			//0x68
+	D2DrlgVertexStrc pVertices[24];			//0x80
+	int32_t nVertices;							//0x260
+	D2DrlgOrthStrc* pRoomData;				//0x264
+};
 
 //D2Common.0x6FD7DC20
 int __fastcall DRLGOUTDOORS_GetOutLinkVisFlag(D2DrlgLevelStrc* pLevel, D2DrlgVertexStrc* pDrlgVertex);
