+++ source/D2Common/include/Drlg/D2DrlgPreset.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "CommonDefinitions.h"
+#include "D2DrlgDrlgGrid.h"
 
 struct D2UnkFileStrc
 {
@@ -8,6 +9,48 @@ struct D2UnkFileStrc
 	long field_104;						//0x104
 	D2DrlgFileStrc* pFile;				//0x108
 	D2UnkFileStrc* pNext;				//0x10C
+};
+
+struct D2DrlgMapStrc
+{
+	int32_t nLevelPrest;						//0x00
+	int32_t nPickedFile;						//0x04
+	D2LvlPrestTxt* pLvlPrestTxtRecord;		//0x08
+	D2DrlgFileStrc* pFile;					//0x0C
+	D2DrlgCoordStrc pDrlgCoord;				//0x10
+	BOOL bHasInfo;							//0x20
+	D2DrlgGridStrc pMapGrid;				//0x24
+	D2PresetUnitStrc* pPresetUnit;			//0x38
+	BOOL bInited;							//0x3C
+	int32_t nPops;								//0x40
+	int32_t* pPopsIndex;						//0x44
+	int32_t* pPopsSubIndex;						//0x48
+	int32_t* pPopsOrientation;					//0x4C
+	D2DrlgCoordStrc* pPopsLocation;			//0x50
+	D2DrlgMapStrc* pNext;					//0x54
+};
+
+struct D2DrlgPresetRoomStrc
+{
+	int32_t nLevelPrest;						//0x00
+	int32_t nPickedFile;						//0x04
+	D2DrlgMapStrc* pMap;					//0x08
+	union
+	{
+		struct
+		{
+			uint8_t nFlags;					//0x0C
+			uint8_t unk0x0D[3];				//0x0D
+		};
+		uint32_t dwFlags;						//0x0C
+	};
+	D2DrlgGridStrc pWallGrid[4];			//0x10
+	D2DrlgGridStrc pOrientationGrid[4];		//0x60
+	D2DrlgGridStrc pFloorGrid[2];			//0xB0
+	D2DrlgGridStrc pCellGrid;				//0xD8
+	D2DrlgGridStrc* pMazeGrid;				//0xEC
+	D2CoordStrc* pTombStoneTiles;			//0xF0
+	int32_t nTombStoneTiles;					//0xF4
 };
 
 //D2Common.0x6FD859A0 (#11222)
