+++ source/D2Common/src/D2Seed.cpp
@@ -1,7 +1,7 @@
 #include "D2Seed.h"
 
 #include <ctime>
-
+#include <CommonDefinitions.h>
 
 //D2Common.0x6FDA5260 (#10916)
 void __stdcall SEED_Return()
