+++ source/D2Common/src/DataTbls/AnimTbls.cpp
@@ -1,6 +1,7 @@
 #include "D2DataTbls.h"
 
 #include "D2Composit.h"
+#include <Units/Units.h>
 
 //TODO: Find names
