+++ source/D2Common/src/DataTbls/DataTbls.cpp
@@ -1,6 +1,8 @@
 #include "D2DataTbls.h"
 #include <cstdio>
 #include <D2Lang.h>
+#include <D2Math.h>
+#include <Units/Units.h>
 
 D2ArenaTxt* gpArenaTxtTable;
 D2CharTemplateTxt* gpCharTemplateTxtTable;
