+++ source/D2Common/src/DataTbls/TokenTbls.cpp
@@ -5,1311 +5,6 @@
 #include "Units/Units.h"
 
 
-//D2Common.0x6FDDE6A8
-D2MonSeqTxt stru_6FDDE6A8[13] =
-{
-	{ 0, PLRMODE_ATTACK1, 0, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 1, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 2, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 3, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 4, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 5, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 6, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 7, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 8, 0, 1 },
-	{ 0, PLRMODE_ATTACK1, 9, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 10, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 11, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 12, 0, 0 },
-};
-
-//D2Common.0x6FDDE6F8
-D2MonSeqTxt stru_6FDDE6F8[14] =
-{
-	{ 0, PLRMODE_ATTACK1, 0, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 1, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 2, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 3, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 4, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 5, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 6, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 7, 0, 1 },
-	{ 0, PLRMODE_ATTACK1, 8, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 9, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 10, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 11, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 12, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 13, 0, 0 },
-};
-
-//D2Common.0x6FDDE750
-D2MonSeqTxt stru_6FDDE750[18] =
-{
-	{ 0, PLRMODE_ATTACK1, 5, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 6, 0, 3 },
-	{ 0, PLRMODE_ATTACK1, 8, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 9, 0, 1 },
-	{ 0, PLRMODE_ATTACK1, 10, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 11, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 13, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 6, 0, 3 },
-	{ 0, PLRMODE_ATTACK2, 8, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 9, 0, 1 },
-	{ 0, PLRMODE_ATTACK1, 10, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 11, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 13, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 6, 0, 3 },
-	{ 0, PLRMODE_ATTACK2, 8, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 9, 0, 1 },
-	{ 0, PLRMODE_ATTACK2, 10, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 13, 0, 0 },
-};
-
-//D2Common.0x6FDDE7C0
-D2MonSeqTxt stru_6FDDE7C0[21] =
-{
-	{ 0, PLRMODE_ATTACK1, 2, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 7, 0, 3 },
-	{ 0, PLRMODE_ATTACK1, 9, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 10, 0, 1 },
-	{ 0, PLRMODE_ATTACK1, 12, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 13, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 15, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 4, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 6, 0, 3 },
-	{ 0, PLRMODE_ATTACK2, 9, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 10, 0, 1 },
-	{ 0, PLRMODE_ATTACK1, 12, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 13, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 15, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 4, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 6, 0, 3 },
-	{ 0, PLRMODE_ATTACK2, 9, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 10, 0, 1 },
-	{ 0, PLRMODE_ATTACK2, 11, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 13, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 15, 0, 0 },
-};
-
-//D2Common.0x6FDDE840
-D2UnkSequenceStrc stru_6FDDE840 =
-{
-	stru_6FDDE6A8, 13, 13,
-	stru_6FDDE750, 18, 18,
-	stru_6FDDE7C0, 21, 21,
-};
-
-//D2Common.0x6FDDE8E8
-D2MonSeqTxt stru_6FDDE8E8[7] =
-{
-	{ 0, PLRMODE_ATTACK2, 0, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 2, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 4, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 7, 0, 1 },
-	{ 0, PLRMODE_ATTACK2, 8, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 10, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 13, 0, 0 },
-};
-
-//D2Common.0x6FDDE914
-D2MonSeqTxt stru_6FDDE914[8] =
-{
-	{ 0, PLRMODE_ATTACK1, 0, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 2, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 4, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 6, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 8, 0, 1 },
-	{ 0, PLRMODE_ATTACK1, 10, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 13, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 15, 0, 0 },
-};
-
-//D2Common.0x6FDDE948
-D2UnkSequenceStrc stru_6FDDE948 =
-{
-	stru_6FDDE6F8, 14, 14,
-	0, 0, 0,
-	0, 0, 0,
-	stru_6FDDE8E8, 7, 7,
-	stru_6FDDE914, 8, 8,
-};
-
-//D2Common.0x6FDDE9F0
-D2MonSeqTxt stru_6FDDE9F0[16] =
-{
-	{ 0, PLRMODE_SPECIAL1, 3, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 4, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 5, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 6, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 7, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 8, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 10, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 1, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 2, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 4, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 6, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 7, 0, 1 },
-	{ 0, PLRMODE_ATTACK2, 8, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 10, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 11, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 13, 0, 0 },
-};
-
-//D2Common.0x6FDDEA50
-D2UnkSequenceStrc stru_6FDDEA50 =
-{
-	stru_6FDDE6F8, 14, 14,
-	0, 0, 0,
-	0, 0, 0,
-	stru_6FDDE9F0, 16, 16,
-};
-
-//D2Common.0x6FDDEAF8
-D2MonSeqTxt stru_6FDDEAF8[15] =
-{
-	{ 0, PLRMODE_RUN, 0, 0, 1 },
-	{ 0, PLRMODE_RUN, 1, 0, 1 },
-	{ 0, PLRMODE_RUN, 2, 0, 1 },
-	{ 0, PLRMODE_RUN, 3, 0, 1 },
-	{ 0, PLRMODE_RUN, 4, 0, 1 },
-	{ 0, PLRMODE_RUN, 5, 0, 1 },
-	{ 0, PLRMODE_RUN, 6, 0, 1 },
-	{ 0, PLRMODE_RUN, 7, 0, 1 },
-	{ 0, PLRMODE_ATTACK1, 1, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 4, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 5, 0, 3 },
-	{ 0, PLRMODE_ATTACK1, 6, 0, 1 },
-	{ 0, PLRMODE_ATTACK1, 8, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 10, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 12, 0, 0 },
-};
-
-//D2Common.0x6FDDEB58
-D2UnkSequenceStrc stru_6FDDEB58 =
-{
-	stru_6FDDEAF8, 15, 15,
-	stru_6FDDEAF8, 15, 15,
-	stru_6FDDEAF8, 15, 15,
-	stru_6FDDEAF8, 15, 15,
-	stru_6FDDEAF8, 15, 15,
-	0, 0, 0,
-	0, 0, 0,
-	stru_6FDDEAF8, 15, 15,
-};
-
-//D2Common.0x6FDDEC00
-D2MonSeqTxt stru_6FDDEC00[6] =
-{
-	{ 0, PLRMODE_ATTACK2, 5, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 6, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 7, 0, 1 },
-	{ 0, PLRMODE_ATTACK2, 8, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 10, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 12, 0, 0 },
-};
-
-//D2Common.0x6FDDEC28
-D2UnkSequenceStrc stru_6FDDEC28 =
-{
-	stru_6FDDE6F8, 14, 14,
-	0, 0, 0,
-	0, 0, 0,
-	stru_6FDDEC00, 6, 6,
-};
-
-//D2Common.0x6FDDECD0
-D2MonSeqTxt stru_6FDDECD0[15] =
-{
-	{ 0, PLRMODE_CAST, 0, 0, 0 },
-	{ 0, PLRMODE_CAST, 1, 0, 0 },
-	{ 0, PLRMODE_CAST, 2, 0, 0 },
-	{ 0, PLRMODE_CAST, 3, 0, 0 },
-	{ 0, PLRMODE_CAST, 4, 0, 0 },
-	{ 0, PLRMODE_CAST, 5, 0, 0 },
-	{ 0, PLRMODE_CAST, 6, 0, 0 },
-	{ 0, PLRMODE_CAST, 7, 0, 0 },
-	{ 0, PLRMODE_CAST, 8, 0, 0 },
-	{ 0, PLRMODE_CAST, 9, 0, 0 },
-	{ 0, PLRMODE_CAST, 9, 0, 1 },
-	{ 0, PLRMODE_CAST, 10, 0, 0 },
-	{ 0, PLRMODE_CAST, 11, 0, 0 },
-	{ 0, PLRMODE_CAST, 12, 0, 0 },
-	{ 0, PLRMODE_CAST, 13, 0, 0 },
-};
-
-//D2Common.0x6FDDED30
-D2UnkSequenceStrc stru_6FDDED30 =
-{
-	stru_6FDDECD0, 15, 15,
-	stru_6FDDECD0, 15, 15,
-	stru_6FDDECD0, 15, 15,
-	stru_6FDDECD0, 15, 15,
-	stru_6FDDECD0, 15, 15,
-	stru_6FDDECD0, 15, 15,
-	stru_6FDDECD0, 15, 15,
-	stru_6FDDECD0, 15, 15,
-	stru_6FDDECD0, 15, 15,
-	stru_6FDDECD0, 15, 15,
-	stru_6FDDECD0, 15, 15,
-	stru_6FDDECD0, 15, 15,
-	stru_6FDDECD0, 15, 15,
-	stru_6FDDECD0, 15, 15,
-};
-
-//D2Common.0x6FDDEDD8
-D2MonSeqTxt stru_6FDDEDD8[13] =
-{
-	{ 0, PLRMODE_ATTACK1, 0, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 1, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 2, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 3, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 4, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 6, 0, 1 },
-	{ 0, PLRMODE_ATTACK1, 7, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 8, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 9, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 10, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 11, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 12, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 13, 0, 0 },
-};
-
-//D2Common.0x6FDDEE28
-D2MonSeqTxt stru_6FDDEE28[20] =
-{
-	{ 0, PLRMODE_ATTACK1, 0, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 1, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 2, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 3, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 4, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 5, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 6, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 7, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 8, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 9, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 10, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 11, 0, 1 },
-	{ 0, PLRMODE_ATTACK1, 12, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 13, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 14, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 15, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 16, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 17, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 18, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 19, 0, 0 },
-};
-
-//D2Common.0x6FDDEEA0
-D2UnkSequenceStrc stru_6FDDEEA0 =
-{
-	stru_6FDDE6A8, 13, 13,
-	0, 0, 0,
-	0, 0, 0,
-	0, 0, 0,
-	0, 0, 0,
-	stru_6FDDEDD8, 13, 13,
-	stru_6FDDEE28, 20, 20,
-};
-
-//D2Common.0x6FDDEF48
-D2MonSeqTxt stru_6FDDEF48[21] =
-{
-	{ 0, PLRMODE_ATTACK1, 0, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 1, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 1, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 1, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 2, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 2, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 2, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 3, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 3, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 4, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 4, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 5, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 6, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 7, 0, 1 },
-	{ 0, PLRMODE_ATTACK1, 8, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 9, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 10, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 11, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 12, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 13, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 14, 0, 0 },
-};
-
-//D2Common.0x6FDDEFC8
-D2MonSeqTxt stru_6FDDEFC8[24] =
-{
-	{ 0, PLRMODE_ATTACK1, 0, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 1, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 1, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 1, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 2, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 2, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 2, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 3, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 3, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 4, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 4, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 5, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 6, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 7, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 8, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 9, 0, 1 },
-	{ 0, PLRMODE_ATTACK1, 10, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 11, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 12, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 13, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 14, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 15, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 16, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 17, 0, 0 },
-};
-
-//D2Common.0x6FDDF058
-D2UnkSequenceStrc stru_6FDDF058 =
-{
-	stru_6FDDE6A8, 13, 13,
-	stru_6FDDEF48, 21, 21,
-	stru_6FDDEFC8, 24, 24,
-};
-
-//D2Common.0x6FDDF100
-D2MonSeqTxt stru_6FDDF100[16] =
-{
-	{ 0, PLRMODE_ATTACK1, 0, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 1, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 2, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 3, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 4, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 5, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 6, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 6, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 8, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 8, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 9, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 9, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 12, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 12, 0, 1 },
-	{ 0, PLRMODE_ATTACK1, 13, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 14, 0, 0 },
-};
-
-//D2Common.0x6FDDF160
-D2MonSeqTxt stru_6FDDF160[16] =
-{
-	{ 0, PLRMODE_ATTACK1, 0, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 1, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 2, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 2, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 6, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 6, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 9, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 9, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 12, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 10, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 14, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 14, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 15, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 16, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 17, 0, 0 },
-	{ 0, 0, 0, 0, 0 },
-};
-
-//D2Common.0x6FDDF1C0
-D2UnkSequenceStrc stru_6FDDF1C0 =
-{
-	stru_6FDDE6A8, 13, 13,
-	stru_6FDDF100, 16, 16,
-	stru_6FDDF160, 16, 16,
-};
-
-//D2Common.0x6FDDF268
-D2MonSeqTxt stru_6FDDF268[8] =
-{
-	{ 0, PLRMODE_ATTACK1, 0, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 1, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 2, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 3, 0, 1 },
-	{ 0, PLRMODE_ATTACK1, 3, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 4, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 5, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 6, 0, 1 },
-};
-
-//D2Common.0x6FDDF298
-D2UnkSequenceStrc stru_6FDDF298 =
-{
-	stru_6FDDF268, 8, 8,
-	stru_6FDDF268, 8, 8,
-	stru_6FDDF268, 8, 8,
-	stru_6FDDF268, 8, 8,
-	stru_6FDDF268, 8, 8,
-	0, 0, 0,
-	0, 0, 0,
-	stru_6FDDF268, 8, 8,
-	stru_6FDDF268, 8, 8,
-	stru_6FDDF268, 8, 8,
-	stru_6FDDF268, 8, 8,
-	stru_6FDDF268, 8, 8,
-	stru_6FDDF268, 8, 8,
-	stru_6FDDF268, 8, 8,
-};
-
-//D2Common.0x6FDDF340
-D2MonSeqTxt stru_6FDDF340[17] =
-{
-	{ 0, PLRMODE_ATTACK1, 1, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 2, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 4, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 6, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 7, 0, 1 },
-	{ 0, PLRMODE_ATTACK1, 9, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 11, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 13, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 14, 0, 0 },
-	{ 0, PLRMODE_SPECIAL3, 2, 0, 1 },
-	{ 0, PLRMODE_SPECIAL3, 4, 0, 0 },
-	{ 0, PLRMODE_SPECIAL3, 5, 0, 0 },
-	{ 0, PLRMODE_SPECIAL3, 6, 0, 0 },
-	{ 0, PLRMODE_SPECIAL3, 8, 0, 0 },
-	{ 0, PLRMODE_SPECIAL3, 9, 0, 0 },
-	{ 0, PLRMODE_SPECIAL3, 10, 0, 0 },
-	{ 0, PLRMODE_SPECIAL3, 11, 0, 0 },
-};
-
-//D2Common.0x6FDDF3A8
-D2UnkSequenceStrc stru_6FDDF3A8 =
-{
-	stru_6FDDF340, 17, 17,
-	stru_6FDDF340, 17, 17,
-	stru_6FDDF340, 17, 17,
-	stru_6FDDF340, 17, 17,
-	stru_6FDDF340, 17, 17,
-	stru_6FDDF340, 17, 17,
-	stru_6FDDF340, 17, 17,
-	stru_6FDDF340, 17, 17,
-	stru_6FDDF340, 17, 17,
-	stru_6FDDF340, 17, 17,
-	stru_6FDDF340, 17, 17,
-	stru_6FDDF340, 17, 17,
-	stru_6FDDF340, 17, 17,
-	stru_6FDDF340, 17, 17,
-};
-
-//D2Common.0x6FDDF450
-D2MonSeqTxt stru_6FDDF450[19] =
-{
-	{ 0, PLRMODE_CAST, 0, 0, 0 },
-	{ 0, PLRMODE_CAST, 1, 0, 0 },
-	{ 0, PLRMODE_CAST, 3, 0, 0 },
-	{ 0, PLRMODE_CAST, 4, 0, 0 },
-	{ 0, PLRMODE_CAST, 5, 0, 0 },
-	{ 0, PLRMODE_CAST, 7, 0, 0 },
-	{ 0, PLRMODE_CAST, 8, 0, 0 },
-	{ 0, PLRMODE_CAST, 9, 0, 1 },
-	{ 0, PLRMODE_CAST, 9, 0, 0 },
-	{ 0, PLRMODE_CAST, 9, 0, 0 },
-	{ 0, PLRMODE_CAST, 9, 0, 0 },
-	{ 0, PLRMODE_CAST, 10, 0, 0 },
-	{ 0, PLRMODE_CAST, 9, 0, 0 },
-	{ 0, PLRMODE_CAST, 9, 0, 0 },
-	{ 0, PLRMODE_CAST, 9, 0, 0 },
-	{ 0, PLRMODE_CAST, 10, 0, 0 },
-	{ 0, PLRMODE_CAST, 11, 0, 0 },
-	{ 0, PLRMODE_CAST, 12, 0, 0 },
-	{ 0, PLRMODE_CAST, 13, 0, 0 },
-};
-
-//D2Common.0x6FDDF4C8
-D2UnkSequenceStrc stru_6FDDF4C8 =
-{
-	stru_6FDDF450, 19, 19,
-	stru_6FDDF450, 19, 19,
-	stru_6FDDF450, 19, 19,
-	stru_6FDDF450, 19, 19,
-	stru_6FDDF450, 19, 19,
-	stru_6FDDF450, 19, 19,
-	stru_6FDDF450, 19, 19,
-	stru_6FDDF450, 19, 19,
-	stru_6FDDF450, 19, 19,
-	stru_6FDDF450, 19, 19,
-	stru_6FDDF450, 19, 19,
-	stru_6FDDF450, 19, 19,
-	stru_6FDDF450, 19, 19,
-	stru_6FDDF450, 19, 19,
-};
-
-//D2Common.0x6FDDF570
-D2MonSeqTxt stru_6FDDF570[15] =
-{
-	{ 0, PLRMODE_SPECIAL1, 0, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 1, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 2, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 3, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 4, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 5, 0, 1 },
-	{ 0, PLRMODE_SPECIAL1, 6, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 7, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 8, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 9, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 10, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 11, 0, 1 },
-	{ 0, PLRMODE_SPECIAL1, 12, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 13, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 14, 0, 1 },
-};
-
-//D2Common.0x6FDDF5D0
-D2UnkSequenceStrc stru_6FDDF5D0 =
-{
-	stru_6FDDF570, 15, 15,
-	stru_6FDDF570, 15, 15,
-	stru_6FDDF570, 15, 15,
-	stru_6FDDF570, 15, 15,
-	stru_6FDDF570, 15, 15,
-	stru_6FDDF570, 15, 15,
-	stru_6FDDF570, 15, 15,
-	stru_6FDDF570, 15, 15,
-	stru_6FDDF570, 15, 15,
-	stru_6FDDF570, 15, 15,
-	stru_6FDDF570, 15, 15,
-	stru_6FDDF570, 15, 15,
-	stru_6FDDF570, 15, 15,
-	stru_6FDDF570, 15, 15,
-};
-
-//D2Common.0x6FDDF678
-D2MonSeqTxt stru_6FDDF678[22] =
-{
-	{ 0, PLRMODE_SPECIAL1, 0, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 1, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 2, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 3, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 4, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 5, 0, 1 },
-	{ 0, PLRMODE_SPECIAL1, 6, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 7, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 8, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 9, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 10, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 11, 0, 1 },
-	{ 0, PLRMODE_SPECIAL1, 12, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 13, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 14, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 5, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 6, 0, 1 },
-	{ 0, PLRMODE_ATTACK1, 7, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 8, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 9, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 10, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 11, 0, 0 },
-};
-
-//D2Common.0x6FDDF700
-D2MonSeqTxt stru_6FDDF700[25] =
-{
-	{ 0, PLRMODE_SPECIAL1, 0, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 1, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 2, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 3, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 4, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 5, 0, 1 },
-	{ 0, PLRMODE_SPECIAL1, 6, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 7, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 8, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 9, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 10, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 11, 0, 1 },
-	{ 0, PLRMODE_SPECIAL1, 12, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 13, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 14, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 6, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 7, 0, 1 },
-	{ 0, PLRMODE_ATTACK1, 8, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 9, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 10, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 11, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 12, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 13, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 14, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 15, 0, 0 },
-};
-
-//D2Common.0x6FDDF798
-D2MonSeqTxt stru_6FDDF798[28] =
-{
-	{ 0, PLRMODE_SPECIAL1, 0, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 1, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 2, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 3, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 4, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 5, 0, 1 },
-	{ 0, PLRMODE_SPECIAL1, 6, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 7, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 8, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 9, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 10, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 11, 0, 1 },
-	{ 0, PLRMODE_SPECIAL1, 12, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 13, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 14, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 6, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 7, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 8, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 9, 0, 1 },
-	{ 0, PLRMODE_ATTACK1, 10, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 11, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 12, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 13, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 14, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 15, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 16, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 17, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 18, 0, 0 },
-};
-
-//D2Common.0x6FDDF840
-D2MonSeqTxt stru_6FDDF840[27] =
-{
-	{ 0, PLRMODE_SPECIAL1, 0, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 1, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 2, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 3, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 4, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 5, 0, 1 },
-	{ 0, PLRMODE_SPECIAL1, 6, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 7, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 8, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 9, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 10, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 11, 0, 1 },
-	{ 0, PLRMODE_SPECIAL1, 12, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 13, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 14, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 6, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 7, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 8, 0, 1 },
-	{ 0, PLRMODE_ATTACK1, 9, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 10, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 11, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 12, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 13, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 14, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 15, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 16, 0, 0 },
-	{ 0, PLRMODE_ATTACK1, 17, 0, 0 },
-};
-
-//D2Common.0x6FDDF8E8
-D2UnkSequenceStrc stru_6FDDF8E8 =
-{
-	stru_6FDDF678, 22, 22,
-	stru_6FDDF700, 25, 25,
-	stru_6FDDF798, 28, 28,
-	stru_6FDDF700, 25, 25,
-	stru_6FDDF840, 27, 27,
-	0, 0, 0,
-	0, 0, 0,
-	stru_6FDDF798, 28, 28,
-	stru_6FDDF678, 22, 22,
-	stru_6FDDF678, 22, 22,
-	stru_6FDDF678, 22, 22,
-	stru_6FDDF678, 22, 22,
-	stru_6FDDF678, 22, 22,
-	stru_6FDDF678, 22, 22,
-};
-
-//D2Common.0x6FDDF990
-D2MonSeqTxt stru_6FDDF990[12] =
-{
-	{ 0, PLRMODE_THROW, 2, 0, 0 },
-	{ 0, PLRMODE_THROW, 3, 0, 0 },
-	{ 0, PLRMODE_THROW, 4, 0, 0 },
-	{ 0, PLRMODE_THROW, 6, 0, 0 },
-	{ 0, PLRMODE_THROW, 7, 0, 0 },
-	{ 0, PLRMODE_THROW, 8, 0, 1 },
-	{ 0, PLRMODE_SPECIAL3, 2, 0, 0 },
-	{ 0, PLRMODE_SPECIAL3, 4, 0, 0 },
-	{ 0, PLRMODE_SPECIAL3, 5, 0, 0 },
-	{ 0, PLRMODE_SPECIAL3, 7, 0, 1 },
-	{ 0, PLRMODE_SPECIAL3, 8, 0, 0 },
-	{ 0, PLRMODE_SPECIAL3, 10, 0, 0 },
-};
-
-//D2Common.0x6FDDF9D8
-D2UnkSequenceStrc stru_6FDDF9D8 =
-{
-	stru_6FDDF990, 12, 12,
-	stru_6FDDF990, 12, 12,
-	stru_6FDDF990, 12, 12,
-	stru_6FDDF990, 12, 12,
-	stru_6FDDF990, 12, 12,
-	stru_6FDDF990, 12, 12,
-	stru_6FDDF990, 12, 12,
-	stru_6FDDF990, 12, 12,
-	stru_6FDDF990, 12, 12,
-	stru_6FDDF990, 12, 12,
-	stru_6FDDF990, 12, 12,
-	stru_6FDDF990, 12, 12,
-	stru_6FDDF990, 12, 12,
-	stru_6FDDF990, 12, 12,
-};
-
-//D2Common.0x6FDDFA80
-D2MonSeqTxt stru_6FDDFA80[12] =
-{
-	{ 0, PLRMODE_ATTACK2, 0, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 1, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 2, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 3, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 4, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 5, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 6, 0, 1 },
-	{ 0, PLRMODE_ATTACK2, 7, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 8, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 9, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 10, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 11, 0, 0 },
-};
-
-//D2Common.0x6FDDFAC8
-D2MonSeqTxt stru_6FDDFAC8[16] =
-{
-	{ 0, PLRMODE_ATTACK2, 0, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 1, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 2, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 3, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 4, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 5, 0, 0 },
-	{ 0, PLRMODE_ATTACK2, 6, 0, 1 },
-	{ 0, PLRMODE_SPECIAL4, 3, 0, 0 },
-	{ 0, PLRMODE_SPECIAL4, 4, 0, 0 },
-	{ 0, PLRMODE_SPECIAL4, 5, 0, 0 },
-	{ 0, PLRMODE_SPECIAL4, 6, 0, 1 },
-	{ 0, PLRMODE_SPECIAL4, 7, 0, 0 },
-	{ 0, PLRMODE_SPECIAL4, 8, 0, 0 },
-	{ 0, PLRMODE_SPECIAL4, 9, 0, 0 },
-	{ 0, PLRMODE_SPECIAL4, 10, 0, 0 },
-	{ 0, PLRMODE_SPECIAL4, 11, 0, 0 },
-};
-
-//D2Common.0x6FDDFB28
-D2UnkSequenceStrc stru_6FDDFB28 =
-{
-	stru_6FDDFA80, 12, 12,
-	0, 0, 0,
-	0, 0, 0,
-	0, 0, 0,
-	0, 0, 0,
-	0, 0, 0,
-	0, 0, 0,
-	0, 0, 0,
-	0, 0, 0,
-	0, 0, 0,
-	0, 0, 0,
-	0, 0, 0,
-	stru_6FDDFA80, 12, 12,
-	stru_6FDDFAC8, 16, 16,
-};
-
-//D2Common.0x6FDDFBD0
-D2MonSeqTxt stru_6FDDFBD0[19] =
-{
-	{ 0, PLRMODE_CAST, 0, 0, 0 },
-	{ 0, PLRMODE_CAST, 1, 0, 0 },
-	{ 0, PLRMODE_CAST, 2, 0, 0 },
-	{ 0, PLRMODE_CAST, 3, 0, 0 },
-	{ 0, PLRMODE_CAST, 4, 0, 0 },
-	{ 0, PLRMODE_CAST, 5, 0, 0 },
-	{ 0, PLRMODE_CAST, 6, 0, 0 },
-	{ 0, PLRMODE_CAST, 7, 0, 1 },
-	{ 0, PLRMODE_CAST, 8, 0, 0 },
-	{ 0, PLRMODE_CAST, 9, 0, 0 },
-	{ 0, PLRMODE_CAST, 10, 0, 0 },
-	{ 0, PLRMODE_CAST, 11, 0, 1 },
-	{ 0, PLRMODE_CAST, 13, 0, 0 },
-	{ 0, PLRMODE_CAST, 14, 0, 0 },
-	{ 0, PLRMODE_CAST, 15, 0, 0 },
-	{ 0, PLRMODE_CAST, 16, 0, 0 },
-	{ 0, PLRMODE_CAST, 17, 0, 0 },
-	{ 0, PLRMODE_CAST, 18, 0, 1 },
-	{ 0, 0, 0, 0, 0 },
-};
-
-//D2Common.0x6FDDFC48
-D2UnkSequenceStrc stru_6FDDFC48 =
-{
-	stru_6FDDFBD0, 19, 19,
-	stru_6FDDFBD0, 19, 19,
-	stru_6FDDFBD0, 19, 19,
-	stru_6FDDFBD0, 19, 19,
-	stru_6FDDFBD0, 19, 19,
-	stru_6FDDFBD0, 19, 19,
-	stru_6FDDFBD0, 19, 19,
-	stru_6FDDFBD0, 19, 19,
-	0, 0, 0,
-	0, 0, 0,
-	0, 0, 0,
-	0, 0, 0,
-	stru_6FDDFBD0, 19, 19,
-	stru_6FDDFBD0, 19, 19,
-};
-
-//D2Common.0x6FDDFCF0
-D2MonSeqTxt stru_6FDDFCF0[20] =
-{
-	{ 0, PLRMODE_KICK, 0, 0, 0 },
-	{ 0, PLRMODE_KICK, 1, 0, 0 },
-	{ 0, PLRMODE_KICK, 2, 0, 0 },
-	{ 0, PLRMODE_KICK, 3, 0, 0 },
-	{ 0, PLRMODE_KICK, 4, 0, 1 },
-	{ 0, PLRMODE_KICK, 5, 0, 0 },
-	{ 0, PLRMODE_KICK, 2, 0, 0 },
-	{ 0, PLRMODE_KICK, 3, 0, 0 },
-	{ 0, PLRMODE_KICK, 4, 0, 1 },
-	{ 0, PLRMODE_KICK, 5, 0, 0 },
-	{ 0, PLRMODE_KICK, 2, 0, 0 },
-	{ 0, PLRMODE_KICK, 3, 0, 0 },
-	{ 0, PLRMODE_KICK, 4, 0, 0 },
-	{ 0, PLRMODE_KICK, 5, 0, 1 },
-	{ 0, PLRMODE_KICK, 6, 0, 0 },
-	{ 0, PLRMODE_KICK, 7, 0, 0 },
-	{ 0, PLRMODE_KICK, 8, 0, 0 },
-	{ 0, PLRMODE_KICK, 9, 0, 0 },
-	{ 0, PLRMODE_KICK, 10, 0, 0 },
-	{ 0, PLRMODE_KICK, 11, 0, 0 },
-};
-
-//D2Common.0x6FDDFD68
-D2UnkSequenceStrc stru_6FDDFD68 =
-{
-	stru_6FDDFCF0, 20, 20,
-	stru_6FDDFCF0, 20, 20,
-	stru_6FDDFCF0, 20, 20,
-	stru_6FDDFCF0, 20, 20,
-	stru_6FDDFCF0, 20, 20,
-	stru_6FDDFCF0, 20, 20,
-	stru_6FDDFCF0, 20, 20,
-	stru_6FDDFCF0, 20, 20,
-	0, 0, 0,
-	0, 0, 0,
-	0, 0, 0,
-	0, 0, 0,
-	stru_6FDDFCF0, 20, 20,
-	stru_6FDDFCF0, 20, 20,
-};
-
-//D2Common.0x6FDDFE10
-D2MonSeqTxt stru_6FDDFE10[15] =
-{
-	{ 0, PLRMODE_SPECIAL1, 0, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 1, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 2, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 3, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 4, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 5, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 6, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 7, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 8, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 9, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 9, 0, 1 },
-	{ 0, PLRMODE_SPECIAL1, 10, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 11, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 12, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 13, 0, 0 },
-};
-
-//D2Common.0x6FDDFE70
-D2UnkSequenceStrc stru_6FDDFE70 =
-{
-	stru_6FDDFE10, 15, 15,
-	stru_6FDDFE10, 15, 15,
-	stru_6FDDFE10, 15, 15,
-	stru_6FDDFE10, 15, 15,
-	stru_6FDDFE10, 15, 15,
-	stru_6FDDFE10, 15, 15,
-	stru_6FDDFE10, 15, 15,
-	stru_6FDDFE10, 15, 15,
-	stru_6FDDFE10, 15, 15,
-	stru_6FDDFE10, 15, 15,
-	stru_6FDDFE10, 15, 15,
-	stru_6FDDFE10, 15, 15,
-	stru_6FDDFE10, 15, 15,
-	stru_6FDDFE10, 15, 15,
-};
-
-//D2Common.0x6FDDFF18
-D2MonSeqTxt stru_6FDDFF18[17] =
-{
-	{ 0, PLRMODE_SPECIAL1, 0, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 1, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 2, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 3, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 4, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 5, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 6, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 7, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 8, 0, 1 },
-	{ 0, PLRMODE_SPECIAL1, 9, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 10, 0, 1 },
-	{ 0, PLRMODE_SPECIAL1, 11, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 12, 0, 1 },
-	{ 0, PLRMODE_SPECIAL1, 12, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 13, 0, 1 },
-	{ 0, PLRMODE_SPECIAL1, 13, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 14, 0, 0 },
-};
-
-//D2Common.0x6FDDFF80
-D2UnkSequenceStrc stru_6FDDFF80 =
-{
-	stru_6FDDFF18, 17, 17,
-	stru_6FDDFF18, 17, 17,
-	stru_6FDDFF18, 17, 17,
-	stru_6FDDFF18, 17, 17,
-	stru_6FDDFF18, 17, 17,
-	stru_6FDDFF18, 17, 17,
-	stru_6FDDFF18, 17, 17,
-	0, 0, 0,
-	0, 0, 0,
-	0, 0, 0,
-	0, 0, 0,
-	0, 0, 0,
-	stru_6FDDFF18, 17, 17,
-	stru_6FDDFF18, 17, 17,
-};
-
-//D2Common.0x6FDE0028
-D2MonSeqTxt stru_6FDE0028[23] =
-{
-	{ 0, PLRMODE_CAST, 0, 0, 0 },
-	{ 0, PLRMODE_CAST, 1, 0, 0 },
-	{ 0, PLRMODE_CAST, 2, 0, 0 },
-	{ 0, PLRMODE_CAST, 3, 0, 0 },
-	{ 0, PLRMODE_CAST, 4, 0, 0 },
-	{ 0, PLRMODE_CAST, 5, 0, 0 },
-	{ 0, PLRMODE_CAST, 6, 0, 0 },
-	{ 0, PLRMODE_CAST, 7, 0, 0 },
-	{ 0, PLRMODE_CAST, 8, 0, 0 },
-	{ 0, PLRMODE_CAST, 9, 0, 1 },
-	{ 0, PLRMODE_KICK, 0, 0, 0 },
-	{ 0, PLRMODE_KICK, 1, 0, 0 },
-	{ 0, PLRMODE_KICK, 2, 0, 0 },
-	{ 0, PLRMODE_KICK, 3, 0, 0 },
-	{ 0, PLRMODE_KICK, 4, 0, 1 },
-	{ 0, PLRMODE_KICK, 5, 0, 0 },
-	{ 0, PLRMODE_KICK, 6, 0, 0 },
-	{ 0, PLRMODE_KICK, 7, 0, 0 },
-	{ 0, PLRMODE_KICK, 8, 0, 0 },
-	{ 0, PLRMODE_KICK, 9, 0, 0 },
-	{ 0, PLRMODE_KICK, 10, 0, 0 },
-	{ 0, PLRMODE_KICK, 11, 0, 0 },
-	{ 0, PLRMODE_KICK, 12, 0, 0 },
-};
-
-//D2Common.0x6FDE00B8
-D2UnkSequenceStrc stru_6FDE00B8 =
-{
-	stru_6FDE0028, 23, 23,
-	stru_6FDE0028, 23, 23,
-	stru_6FDE0028, 23, 23,
-	stru_6FDE0028, 23, 23,
-	stru_6FDE0028, 23, 23,
-	stru_6FDE0028, 23, 23,
-	stru_6FDE0028, 23, 23,
-	stru_6FDE0028, 23, 23,
-	stru_6FDE0028, 23, 23,
-	stru_6FDE0028, 23, 23,
-	stru_6FDE0028, 23, 23,
-	stru_6FDE0028, 23, 23,
-	stru_6FDE0028, 23, 23,
-	stru_6FDE0028, 23, 23,
-};
-
-//D2Common.0x6FDE0160
-D2MonSeqTxt stru_6FDE0160[16] =
-{
-	{ 0, PLRMODE_SPECIAL1, 15, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 14, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 13, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 12, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 11, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 10, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 9, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 8, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 7, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 6, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 5, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 4, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 3, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 2, 0, 1 },
-	{ 0, PLRMODE_SPECIAL1, 1, 0, 0 },
-	{ 0, PLRMODE_SPECIAL1, 0, 0, 0 },
-};
-
-//D2Common.0x6FDE01C0
-D2UnkSequenceStrc stru_6FDE01C0 =
-{
-	stru_6FDE0160, 16, 16,
-	stru_6FDE0160, 16, 16,
-	stru_6FDE0160, 16, 16,
-	stru_6FDE0160, 16, 16,
-	stru_6FDE0160, 16, 16,
-	stru_6FDE0160, 16, 16,
-	stru_6FDE0160, 16, 16,
-	stru_6FDE0160, 16, 16,
-	stru_6FDE0160, 16, 16,
-	stru_6FDE0160, 16, 16,
-	stru_6FDE0160, 16, 16,
-	stru_6FDE0160, 16, 16,
-	stru_6FDE0160, 16, 16,
-	stru_6FDE0160, 16, 16,
-};
-
-//D2Common.0x6FDE0268
-D2MonSeqTxt stru_6FDE0268[19] =
-{
-	{ 0, PLRMODE_CAST, 0, 0, 0 },
-	{ 0, PLRMODE_CAST, 1, 0, 0 },
-	{ 0, PLRMODE_CAST, 2, 0, 0 },
-	{ 0, PLRMODE_CAST, 3, 0, 0 },
-	{ 0, PLRMODE_CAST, 4, 0, 0 },
-	{ 0, PLRMODE_CAST, 5, 0, 0 },
-	{ 0, PLRMODE_CAST, 6, 0, 0 },
-	{ 0, PLRMODE_CAST, 7, 0, 0 },
-	{ 0, PLRMODE_CAST, 8, 0, 0 },
-	{ 0, PLRMODE_CAST, 9, 0, 0 },
-	{ 0, PLRMODE_CAST, 9, 0, 0 },
-	{ 0, PLRMODE_CAST, 9, 0, 0 },
-	{ 0, PLRMODE_CAST, 10, 0, 1 },
-	{ 0, PLRMODE_CAST, 11, 0, 0 },
-	{ 0, PLRMODE_CAST, 12, 0, 0 },
-	{ 0, PLRMODE_CAST, 13, 0, 0 },
-	{ 0, PLRMODE_CAST, 14, 0, 0 },
-	{ 0, PLRMODE_CAST, 15, 0, 0 },
-	{ 0, PLRMODE_CAST, 16, 0, 0 },
-};
-
-//D2Common.0x6FDE02E0
-D2UnkSequenceStrc stru_6FDE02E0 =
-{
-	stru_6FDE0268, 19, 19,
-	stru_6FDE0268, 19, 19,
-	stru_6FDE0268, 19, 19,
-	stru_6FDE0268, 19, 19,
-	stru_6FDE0268, 19, 19,
-	stru_6FDE0268, 19, 19,
-	stru_6FDE0268, 19, 19,
-	stru_6FDE0268, 19, 19,
-	stru_6FDE0268, 19, 19,
-	stru_6FDE0268, 19, 19,
-	stru_6FDE0268, 19, 19,
-	stru_6FDE0268, 19, 19,
-	stru_6FDE0268, 19, 19,
-	stru_6FDE0268, 19, 19,
-};
-
-//D2Common.0x6FDE0388
-D2UnkSequenceStrc* off_6FDE0388[24] =
-{
-	NULL,
-	&stru_6FDDE840,
-	&stru_6FDDE948,
-	&stru_6FDDEA50,
-	&stru_6FDDEB58,
-	&stru_6FDDEC28,
-	&stru_6FDDED30,
-	&stru_6FDDEEA0,
-	&stru_6FDDF058,
-	&stru_6FDDF1C0,
-	&stru_6FDDF298,
-	&stru_6FDDF3A8,
-	&stru_6FDDF4C8,
-	&stru_6FDDF5D0,
-	&stru_6FDDF8E8,
-	&stru_6FDDF9D8,
-	&stru_6FDDFB28,
-	&stru_6FDDFC48,
-	&stru_6FDDFE70,
-	&stru_6FDDFD68,
-	&stru_6FDDFF80,
-	&stru_6FDE00B8,
-	&stru_6FDE01C0,
-	&stru_6FDE02E0,
-};
-
-//D2Common.0x6FDE03E8
-static const int dword_6FDE03E8[14][2] =
-{
-	{ 0, WEAPONCLASS_NONE },
-	{ 1, WEAPONCLASS_1HT },
-	{ 2, WEAPONCLASS_2HT },
-	{ 3, WEAPONCLASS_1HS },
-	{ 4, WEAPONCLASS_2HS },
-	{ 5, WEAPONCLASS_BOW },
-	{ 6, WEAPONCLASS_XBW },
-	{ 7, WEAPONCLASS_STF },
-	{ 8, WEAPONCLASS_1JS },
-	{ 9, WEAPONCLASS_1JT },
-	{ 10, WEAPONCLASS_1SS },
-	{ 11, WEAPONCLASS_1ST },
-	{ 12, WEAPONCLASS_HT1 },
-	{ 13, WEAPONCLASS_HT2 }
-};
-
-
-
-//D2Common.0x6FD727A0 (#10682)
-D2MonSeqTxt* __stdcall DATATBLS_GetMonSeqTxtRecordFromUnit(D2UnitStrc* pUnit)
-{
-	D2SeqRecordStrc* pSeqRecord = DATATBLS_GetSeqRecordFromUnit(pUnit);
-	if (pSeqRecord)
-	{
-		return pSeqRecord->pMonSeqTxtRecord;
-	}
-
-	return NULL;
-}
-
-//D2Common.0x6FD727C0
-D2SeqRecordStrc* __fastcall DATATBLS_GetSeqRecordFromUnit(D2UnitStrc* pUnit)
-{
-	int nSequenceNum = 0;
-	int nWeaponClass = 0;
-	int nUnitType = 0;
-
-	if(D2SkillStrc* pSkill = UNITS_GetUsedSkill(pUnit))
-	{
-		nSequenceNum = SKILLS_GetSeqNumFromSkill(pUnit, pSkill);
-
-		if (nSequenceNum > 0)
-		{
-			if (pUnit)
-			{
-				nUnitType = pUnit->dwUnitType;
-			}
-			else
-			{
-				nUnitType = 6;
-			}
-
-			if (nUnitType != UNIT_PLAYER)
-			{
-				if (nUnitType == UNIT_MONSTER && DATATBLS_GetMonStatsTxtRecord(pUnit->dwClassId))
-				{
-					return DATATBLS_GetMonSeqTableRecord(nSequenceNum);
-				}
-			}
-			else
-			{
-				COMPOSIT_GetWeaponClassId(pUnit, pUnit->pInventory, &nWeaponClass, -1, TRUE);
-
-				int nWClassIndex = -1;
-				for (int i = 0; i < 14; i++)
-				{
-					if (nWeaponClass == dword_6FDE03E8[i][1])
-					{
-						nWClassIndex = dword_6FDE03E8[i][0];
-						break;
-					}
-				}
-
-				D2_ASSERT(nWClassIndex != -1);
-
-				return &off_6FDE0388[nSequenceNum]->unk0x00[nWClassIndex];
-			}
-		}
-	}
-
-	return NULL;
-}
-
-//D2Common.0x6FD728A0 (#10683)
-//TODO: Find a name
-int __stdcall D2Common_10683(D2UnitStrc* pUnit)
-{
-	D2SeqRecordStrc* pSeqRecord = DATATBLS_GetSeqRecordFromUnit(pUnit);
-	if (pSeqRecord)
-	{
-		return (pSeqRecord->unk0x04 << 8);
-	}
-
-	return 0;
-}
-
-//D2Common.0x6FD728C0 (#10684)
-//TODO: Find a name
-int __stdcall D2Common_10684(D2UnitStrc* pUnit)
-{
-	D2SeqRecordStrc* pSeqRecord = DATATBLS_GetSeqRecordFromUnit(pUnit);
-	if (pSeqRecord)
-	{
-		return pSeqRecord->unk0x08;
-	}
-
-	return 0;
-}
-
-//D2Common.0x6FD728E0 (#10685)
-//TODO: Find a name
-void __stdcall D2Common_10685(D2MonSeqTxt* pMonSeqTxt, int a2, int a3, unsigned int* pMode, unsigned int* pFrame, int* pDirection, int* pEvent)
-{
-	D2MonSeqTxt* pMonSeqTxtRecord = NULL;
-	int nCounter = 0;
-
-	if (pMonSeqTxt)
-	{
-		pMonSeqTxtRecord = &pMonSeqTxt[a2 >> 8];
-
-		*pMode = pMonSeqTxtRecord->nMode;
-		*pFrame = pMonSeqTxtRecord->nFrame;
-		*pDirection = pMonSeqTxtRecord->nDir;
-
-		if (a2 == a3)
-		{
-			*pEvent = pMonSeqTxtRecord->nEvent;
-		}
-		else
-		{
-			*pEvent = 0;
-
-			if ((a3 >> 8) + 1 <= (a2 >> 8))
-			{
-				nCounter = 0;
-				do
-				{
-					if (pMonSeqTxt[(a3 >> 8) + 1 + nCounter].nEvent)
-					{
-						*pEvent = pMonSeqTxt[(a3 >> 8) + 1 + nCounter].nEvent;
-					}
-					++nCounter;
-				}
-				while (nCounter < ((a2 >> 8) - (a3 >> 8)));
-			}
-		}
-	}
-	else
-	{
-		*pMode = 0;
-		*pFrame = 0;
-		*pDirection = 0;
-		*pEvent = 0;
-	}
-}
-
-//D2Common.0x6FD72990 (#10686)
-void __stdcall DATATBLS_GetSequenceEvent(D2MonSeqTxt* pMonSeqTxt, int nSeqFrame, int* pEvent)
-{
-	if (pMonSeqTxt)
-	{
-		*pEvent = pMonSeqTxt[nSeqFrame >> 8].nEvent;
-	}
-	else
-	{
-		*pEvent = 0;
-	}
-}
 
 //D2Common.0x6FD729C0
 void __fastcall DATATBLS_LoadPlrType_ModeTxt(void* pMemPool)
@@ -1574,11 +269,4 @@ D2ArmTypeTxt* __stdcall DATATBLS_GetArmTypeTxtRecord(int nId)
 {
 	D2_ASSERT(&sgptDataTables->pArmTypeTxt[nId]);
 	return &sgptDataTables->pArmTypeTxt[nId];
-}
-
-//D2Common.0x6FD733B0 (#10667)
-int __stdcall DATATBLS_UNUSED_Return0(int a1, int a2)
-{
-	FOG_10025("0", __FILE__, __LINE__);
-	return 0;
 }
