+++ source/D2Common/src/DataTbls/TransformTbls.cpp
@@ -1,0 +1,9 @@
+#include <DataTbls/TransformTbls.h>
+#include <Fog.h>
+
+//D2Common.0x6FD733B0 (#10667)
+int __stdcall DATATBLS_UNUSED_Return0(int a1, int a2)
+{
+	D2_UNREACHABLE;
+	return 0;
+}
