+++ source/D2Common/src/Drlg/DrlgOutdoors.cpp
@@ -9,6 +9,7 @@
 #include "Drlg/D2DrlgMaze.h"
 #include "Drlg/D2DrlgOutDesr.h"
 #include "Drlg/D2DrlgOutPlace.h"
+#include "Drlg/D2DrlgOutRoom.h"
 #include "Drlg/D2DrlgOutSiege.h"
 #include "Drlg/D2DrlgOutWild.h"
 #include "Drlg/D2DrlgPreset.h"
