+++ source/D2Common/src/Drlg/DrlgTileSub.cpp
@@ -7,6 +7,8 @@
 #include "Drlg/D2DrlgMaze.h"
 #include "Drlg/D2DrlgPreset.h"
 #include "Drlg/D2DrlgRoomTile.h"
+#include "Drlg/D2DrlgOutdoors.h"
+#include "Drlg/D2DrlgOutRoom.h"
 #include "D2Dungeon.h"
 #include "D2Seed.h"
